### 2024-06-19 #6901 @editorialbot
[PRE REVIEW]: pymnet: A Multilayer-Networks Library for Python
**Submitting author:** @ercco (<a href="http://orcid.org/0000-0003-0258-7776">Tarmo Nurmi</a>)
**Repository:** https://github.com/mnets/pymnet
**Branch with paper.md** (empty if default branch): publication
**Version:** 1.0.0
