### 2024-06-20 50ce624
add dois to references, update bliss paper info
