+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={203--271},
   year={2014},
   publisher={Oxford University Press}
+  doi={10.1093/comnet/cnu016}
 }
 
 @article{kivela2017,
@@ -18,6 +19,7 @@
   pages={198--211},
   year={2017},
   publisher={IEEE}
+  doi={10.1109/TNSE.2017.2753963}
 }
 
 @article{sallmen2022,
@@ -29,6 +31,7 @@
   pages={cnac005},
   year={2022},
   publisher={Oxford University Press}
+  doi={10.1093/comnet/cnac005}
 }
 
 @misc{nurmi2023,
@@ -38,26 +41,39 @@
       eprint={2308.00083},
       archivePrefix={arXiv},
       primaryClass={physics.soc-ph}
+      doi={10.48550/arXiv.2308.00083}
 }
 
 @InProceedings{junttila2011,
   author =    {Tommi Junttila and Petteri Kaski},
   title =     {Conflict Propagation and Component Recursion for Canonical Labeling},
-  booktitle = {Theory and Practice of Algorithms in (Computer) Systems (ICST)},
+  booktitle = {Theory and Practice of Algorithms in (Computer) Systems -- First International {ICST} Conference, {TAPAS} 2011, Rome, Italy, April 18--20, 2011. Proceedings},
   year =      {2011},
-  pages =     {151--162}
+  editor =    {Alberto Marchetti{-}Spaccamela and Michael Segal},
+  volume =    {6595},
+  series =    {Lecture Notes in Computer Science},
+  pages =     {151--162},
+  publisher = {Springer},
+  doi =       {10.1007/978-3-642-19754-3\_16},
 }
 
 @InProceedings{junttila2007,
   author =     {Tommi Junttila and Petteri Kaski},
-  title =      {Engineering an Efficient Canonical Labeling Tool for
-                Large and Sparse Graphs},
-  booktitle =  {{Proceedings of the Ninth Workshop on Algorithm Engineering
+  title =      {Engineering an efficient canonical labeling tool for
+                large and sparse graphs},
+  booktitle =  {Proceedings of the Ninth Workshop on Algorithm Engineering
                 and Experiments and
                 the Fourth Workshop on Analytic Algorithms and
-                Combinatorics}},
+                Combinatorics},
   pages =      {135--149},
-  year =       {2007}
+  year =       {2007},
+  editor =     {David Applegate and Gerth St{\o}lting Brodal and
+                Daniel Panario and Robert Sedgewick},
+  OPTaddress = {New Orleans, LA},
+  OPTmonth =   {January 6},
+  publisher =  {SIAM},
+  OPTpublisher = {Society for Industrial and Applied Mathematics},
+  doi = {10.1137/1.9781611972870.13},
 }
 
 @article{yaverouglu2014,
@@ -69,6 +85,7 @@
   pages={4547},
   year={2014},
   publisher={Nature Publishing Group UK London}
+  doi={10.1038/srep04547}
 }
 
 @article{zhou2020network,
@@ -80,6 +97,7 @@
   pages={1388--1397},
   year={2020},
   publisher={IEEE}
+  doi={10.1109/TEM.2020.3032160}
 }
 
 @article{del2020multiplex,
@@ -90,6 +108,7 @@
   pages={1--23},
   year={2020},
   publisher={Springer}
+  doi={10.1007/s41109-020-00301-2}
 }
 
 @article{cozzo2015structure,
@@ -101,6 +120,7 @@
   pages={073029},
   year={2015},
   publisher={IOP Publishing}
+  doi={10.1088/1367-2630/17/7/073029}
 }
 
[…]
