### 2024-06-20 92b3aff
add missing commas in paper.bib and remove unnecessary ones
