+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={3},
   pages={203--271},
   year={2014},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/comnet/cnu016}
 }
 
@@ -18,7 +18,7 @@
   number={3},
   pages={198--211},
   year={2017},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/TNSE.2017.2753963}
 }
 
@@ -30,7 +30,7 @@
   number={2},
   pages={cnac005},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/comnet/cnac005}
 }
 
@@ -40,7 +40,7 @@
       year={2023},
       eprint={2308.00083},
       archivePrefix={arXiv},
-      primaryClass={physics.soc-ph}
+      primaryClass={physics.soc-ph},
       doi={10.48550/arXiv.2308.00083}
 }
 
@@ -54,7 +54,7 @@
   series =    {Lecture Notes in Computer Science},
   pages =     {151--162},
   publisher = {Springer},
-  doi =       {10.1007/978-3-642-19754-3\_16},
+  doi =       {10.1007/978-3-642-19754-3\_16}
 }
 
 @InProceedings{junttila2007,
@@ -73,7 +73,7 @@
   OPTmonth =   {January 6},
   publisher =  {SIAM},
   OPTpublisher = {Society for Industrial and Applied Mathematics},
-  doi = {10.1137/1.9781611972870.13},
+  doi = {10.1137/1.9781611972870.13}
 }
 
 @article{yaverouglu2014,
@@ -84,7 +84,7 @@
   number={1},
   pages={4547},
   year={2014},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi={10.1038/srep04547}
 }
 
@@ -96,7 +96,7 @@
   number={5},
   pages={1388--1397},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/TEM.2020.3032160}
 }
 
@@ -107,7 +107,7 @@
   volume={5},
   pages={1--23},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/s41109-020-00301-2}
 }
 
@@ -119,7 +119,7 @@
   number={7},
   pages={073029},
   year={2015},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi={10.1088/1367-2630/17/7/073029}
 }
 
@@ -130,7 +130,7 @@
   volume={8},
   pages={e43094},
   year={2019},
-  publisher={eLife Sciences Publications, Ltd}
+  publisher={eLife Sciences Publications, Ltd},
   doi={10.7554/eLife.43094}
 }
 
@@ -142,7 +142,7 @@
   number={8},
   pages={883--901},
   year={2021},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/scan/nsaa069}
 }
 
@@ -153,7 +153,7 @@
   volume={6},
   pages={1--33},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/s41109-021-00429-9}
 }
 
