+++ paper/paper.bib
@@ -54,7 +54,7 @@
   series =    {Lecture Notes in Computer Science},
   pages =     {151--162},
   publisher = {Springer},
-  doi =       {10.1007/978-3-642-19754-3\_16}
+  doi =       {10.1007/978-3-642-19754-3_16}
 }
 
 @InProceedings{junttila2007,
