### 2024-06-23 #6901 @ClaudMor
Dear @danielskatz,

Both @pitmonticone and I would agree to review this package.

In accordance with the Conflict of Interest rules, I must disclose that we are listed as contributors to the software under review. This is due to a [single commit](https://github.com/mnets/pymnet/commit/a6fc958562986036f8e6b80d4ec3db6a129e9687) where we corrected a few typographical errors. Beyond this minor contribution, we have not been involved in the development of this package and have had no other prior, concurrent or subsequent interactions with the authors.

If this does not present a conflict, we are prepared to proceed with the review, given the close similarity of this package to our [previous work](https://doi.org/10.21105/joss.05116).
