### 2024-06-26 #6901 @danielskatz
@leotrs, @nwlandry, @LCB0B - Would one (or even two) of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
