### 2024-06-26 #6930 @editorialbot
[REVIEW]: pymnet: A Python Library for Multilayer Networks
**Submitting author:** @ercco (<a href="http://orcid.org/0000-0003-0258-7776">Tarmo Nurmi</a>)
**Repository:** https://github.com/mnets/pymnet
**Branch with paper.md** (empty if default branch): publication
**Version:** v1.0.0
