### 2024-06-28 mnets/pymnet#22 @pitmonticone
[REVIEW]: Installation instructions in docs/tutorial
In the [docs/tutorial](https://mnets.github.io/pymnet/tutorial/installing.html)
- [x] the mentioned URL to the GitHub repository works, but it is a redirect. Could you update it?
- [x] the mentioned git clone path is outdated and requires a public SSH key. Could you substitute it with the HTTPS version (i.e. `git clone https://github.com/mnets/pymnet.git`)?
