### 2024-06-28 #6930 @pitmonticone
https://github.com/mnets/pymnet/issues/22
