### 2024-06-28 mnets/pymnet#24 @pitmonticone
[REVIEW]: Paper title
The current paper title is “`pymnet`: A Multilayer-Networks Library for Python”.

I believe the hyphen in “Multilayer-Networks” is unnecessary. I suggest removing the hyphen and considering alternative titles such as:
1. “`pymnet`: A Multilayer Networks Library for Python”
2. “`pymnet`: A Python Library for Multilayer Networks”
3. “`pymnet`: A Multilayer Networks Library in Python”
