### 2024-06-29 mnets/pymnet#25 @ClaudMor
[REVIEW] Typo in Tutorial: Network types: Multiplex networks
Dear authors,

In the tutorial section about [multiplex networks](https://mnets.github.io/pymnet/tutorial/networktypes.html#multiplex-networks), after defining the 2-aspect multiplex network `conet` you try to add a layer to its third aspect, which results in an error.

```python
conet = MultiplexNetwork(couplings=['categorical','ordinal'])
conet.add_node('node')
conet.add_layer('a',1)
conet.add_layer('b',1)
conet.add_layer(1,2)
conet.add_layer(2,2)
conet.add_layer(3,3)
```
```bash
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\claud\anaconda3\envs\pymnet_review\Lib\site-packages\pymnet\net.py", line 997, in add_layer
    if layer not in self.slices[aspect]:
                    ~~~~~~~~~~~^^^^^^^^
IndexError: list index out of range
```
