### 2024-06-29 mnets/pymnet#26 @ClaudMor
[REVIEW] Missing documentation about `pymnet.net.MultilayerNode`
Dear authors,

While running your tutorial, I noticed the existence of the class `pymnet.net.MultilayerNode`:

```python
from pymnet import *
net = MultilayerNetwork(aspects=0)
net.add_node(1)
net[1]
```
```nothing
<pymnet.net.MultilayerNode object at 0x000002CD29CAAA50>
```

which exists [here](https://github.com/mnets/pymnet/blob/fc22a879f13fa91f8ebe49186c19553ce19d01a2/pymnet/net.py#L656).

Anyway I could not find documentation for it. Since it exposes public APIs, I'd suggest providing such documentation.

Also, I noticed the use of `str` to name the function that returns the weighted degree of a node. Since the `str` name is historically assigned to strings in Python, I would suggest renaming it to something else.
