### 2024-06-29 mnets/pymnet#27 @ClaudMor
[REVIEW} Strange behaviours
Dear authors,

While reviewing your package, I noticed the following unexpected behaviours:

1. Assigning strings to edges is allowed:

```python
from pymnet import *
net = MultilayerNetwork(aspects=0)
net[1,2] = "hello"
net[1,2]
```
```nothing
"hello"
```

2. Modifying the `MultilayerNetwork` via a `MultilayerNode`:

```python
from pymnet import *
net = MultilayerNetwork(aspects=0)
net.add_node(1)
net[1].mnet.add_node(2)
list(net)
```
```nothing
[1, 2]
```
I understand that these behaviors do not break the software, therefore fixing them is not mandatory for this review. Anyway, for the future, I would like to respectively point out:

1. Disallowing this by e.g. throwing and error;
2. Using [getters and setters](https://docs.python.org/3/library/functions.html#property), or [name mangling](https://www.geeksforgeeks.org/name-mangling-in-python/);
