### 2024-06-29 mnets/pymnet#28 @ClaudMor
[REVIEW] Figure not showing
Dear authors,

If I run your tutorial on [Visualizing Networks](https://mnets.github.io/pymnet/tutorial/visualizing.html), the figure does not show:

```python
from pymnet import *
net = models.er_multilayer(5,2,0.2)
fig = draw(net,show=True)
```
```nothing
C:\Users\claud\anaconda3\envs\pymnet_review\Lib\site-packages\pymnet\visuals\drawbackends\mpl.py:79: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
  plt.show()
```

I tried installing PyQt6 (as suggested [here](https://stackoverflow.com/a/78344937)) to no avail.

Would you have any idea on what could the problem be (perhaps it is related to my machine or environment)?

Environment:

```nothing
(pymnet_review) C:\Users\claud>conda list
# packages in environment at C:\Users\claud\anaconda3\envs\pymnet_review:
#
# Name                    Version                   Build  Channel
bzip2                     1.0.8                h2bbff1b_6
ca-certificates           2024.3.11            haa95532_0
contourpy                 1.2.1                    pypi_0    pypi
cycler                    0.12.1                   pypi_0    pypi
expat                     2.6.2                hd77b12b_0
fonttools                 4.53.0                   pypi_0    pypi
kiwisolver                1.4.5                    pypi_0    pypi
libffi                    3.4.4                hd77b12b_1
matplotlib                3.9.0                    pypi_0    pypi
networkx                  3.3                      pypi_0    pypi
numpy                     2.0.0                    pypi_0    pypi
openssl                   3.0.14               h827c3e9_0
packaging                 24.1                     pypi_0    pypi
pillow                    10.3.0                   pypi_0    pypi
pip                       24.0            py312haa95532_0
pymnet                    0.2.0                    pypi_0    pypi
pyparsing                 3.1.2                    pypi_0    pypi
pyqt6                     6.7.0                    pypi_0    pypi
pyqt6-qt6                 6.7.2                    pypi_0    pypi
pyqt6-sip                 13.6.0                   pypi_0    pypi
python                    3.12.4               h14ffc60_1
python-dateutil           2.9.0.post0              pypi_0    pypi
scipy                     1.13.1                   pypi_0    pypi
setuptools                69.5.1          py312haa95532_0
six                       1.16.0                   pypi_0    pypi
sqlite                    3.45.3               h2bbff1b_0
tk                        8.6.14               h0416ee5_0
tzdata                    2024a                h04d1e81_0
vc                        14.2                 h2eaa2aa_1
vs2015_runtime            14.29.30133          h43f2093_3
wheel                     0.43.0          py312haa95532_0
xz                        5.4.6                h8cc25b3_1
zlib                      1.2.13               h8cc25b3_1
```

OS: windows 10

IDE: vscode
