### 2024-06-29 mnets/pymnet#29 @ClaudMor
[REVIEW] Error when computing automorphism
Dear authors,

While running the tutorial on [automorphisms](https://mnets.github.io/pymnet/tutorial/isomorphisms.html#automorphisms), I got the error:

```python
import pymnet
net_social = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
net_social["Alice", "Bob", "Friends"] = 1
net_social["Alice", "Carol", "Friends"] = 1
net_social["Bob", "Carol", "Friends"] = 1
net_social["Alice", "Bob", "Married"] = 1

net_transport = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
net_transport["Helsinki", "Turku", "Train"] = 1
net_transport["Helsinki", "Tampere", "Train"] = 1
net_transport["Turku", "Tampere", "Train"] = 1
net_transport["Helsinki", "Turku", "Ferry"] = 1

print(pymnet.get_automorphism_generators(net_social, allowed_aspects=[0]))
```
```nothing
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "C:\Users\claud\anaconda3\envs\pymnet_review\Lib\site-packages\pymnet\isomorphisms\__init__.py", line 197, in get_automorphism_generators
    assert len(automorphism_group_generator_backends) > 0, \
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AssertionError: No backends for automorphism generators were imported!
```

I understand that the `bliss-blind` package is needed. Anyway, I could not find it online (also, it should come installed as a dependency on Windows x64 machines, according to the documentation).
