### 2024-06-29 mnets/pymnet#30 @ClaudMor
[REVIEW] Residuals from Python 2
Dear authors,

I noticed that in the tutorial on [Using NetworkX functions](https://mnets.github.io/pymnet/tutorial/nx.html#using-networkx-functions), some residuals from Python 2 are present:

```python
print net
```

```python
print map(nx.number_connected_components,models.er(1000,3*[0.005]).A.values())
```

Moreover the last line, when ported to python 3, does not return `[9, 11, 8]` as reported:

```python
print(map(nx.number_connected_components,models.er(1000,3*[0.005]).A.values()))
```
```nothing
<map object at 0x000001B26309D4B0>
```
