### 2024-06-29 mnets/pymnet#29 @arashbm
Interesting. The package is [definitely available on PyPI](https://pypi.org/project/bliss-bind/) and here on [GitHub](https://github.com/mnets/bliss_bind), but since it is a native extension, it is only set to install automatically as a dependency if we have it pre-built for the specific combination of Python abi version, operating system/libc variant and CPU architecture that you are using. [Here](https://pypi.org/project/bliss-bind/#files) is a list of environment that we have it built for. I can see that we have a combination for AMD64 on Windows with CPython8 to 12.

I think what's happening here is that the logic we use to detect if we should install it as a dependency is faulty on windows. Here is how the dependency is defined at the moment:

```
"bliss-bind >= 0.2.0 ; (sys_platform == 'darwin' or ( (platform_machine == 'x86_64' and sys_platform == 'linux') or (platform_machine == 'amd64' and sys_platform == 'win32') )) and python_version >= '3.8'"
```

Can you please test what are the values that this excerpt prints?

```python
$ python -c "import sys, platform; print(sys.platform, platform.machine())"
```

Also, while we try to resolve this, you can try pip installing `bliss_bind` manually and running the test again. If it fails that gives another data point.

```
$ pip install -U bliss-bind
```
