### 2024-06-29 #6930 @ClaudMor
### Review Issues

- https://github.com/mnets/pymnet/issues/25
- https://github.com/mnets/pymnet/issues/26
- https://github.com/mnets/pymnet/issues/27
- https://github.com/mnets/pymnet/issues/28
- https://github.com/mnets/pymnet/issues/29
- https://github.com/mnets/pymnet/issues/30
