### 2024-07-01 6e72e4b
update installation instructions #22
