+++ doc/tutorial/installing.rst
@@ -7,14 +7,13 @@ The easiest way to install pymnet is via pip:
 
     $ pip install pymnet
 
-We recommend installing into a virtual environment (e.g., managed with virtualenv, conda, or poetry).
+We recommend installing into a virtual environment (managed, e.g., with virtualenv, conda, mamba, or poetry).
 Afterwards, the installation should work out-of-the-box.
 
-Alternatively, you can download the latest source files directly from the repository https://github.com/bolozna/Multilayer-networks-library as a zip file,
-or you can use Git to clone the repository:
+Alternatively, you can download the latest source files directly from our `GitHub Repository <https://github.com/mnets/pymnet>`_ as a zip file,
+or you can use git to clone the repository, for example, via HTTPS:
 
 .. code-block:: none
 
-   $ git clone git@github.com:bolozna/Multilayer-networks-library.git
+   $ git clone https://github.com/mnets/pymnet.git
 
-The current version of the library is written purely in Python, and simply adding the library directory to your Python path is enough to start using it.
