### 2024-07-01 mnets/pymnet#30 @dataspider
Thank you for catching these! I'll make a pass over all tutorials and ensure they are completely ported to python3. While I'm at it, I'll also translate the tutorials into Jupyter Notebooks (this was on my TODO list anyway).
This should also take care of some of the other issues you identified.
