### 2024-07-01 9d2a71e
add MultilayerNode to docs #26
