### 2024-07-01 45e5a8d
fix some MultilayerNode doc issues #26
