+++ pymnet/net.py
@@ -139,14 +139,14 @@ class MultilayerNetwork(object):
         
     #@classmehtod
     def _link_to_nodes(self,link):
-        """Returns the link as tuple of nodes in the graph representing
+        """Return the link as tuple of nodes in the graph representing
         the multislice structure. I.e. when given (i,j,s_1,r_1, ... ,s_d,r_d)
         (i,s_1,...,s_d),(j,r_1,...,r_d) is returned.
         """
         return (link[0],)+link[2::2],(link[1],)+link[3::2]
     #@classmehtod
     def _nodes_to_link(self,node1,node2):
-        """Returns a link when tuple of nodes is given in the graph representing
+        """Return a link when tuple of nodes is given in the graph representing
         the multislice structure. I.e. when given (i,s_1,...,s_d),(j,r_1,...,r_d) 
         (i,j,s_1,r_1, ... ,s_d,r_d) is returned.
         """
@@ -158,7 +158,7 @@ class MultilayerNetwork(object):
         return tuple(l)
     #@classmehtod
     def _short_link_to_link(self,slink):
-        """ Returns a full link for the shortened version of the link. That is,
+        """Return a full link for the shortened version of the link. That is,
         if (i,j,s_1,...,s_d) is given as input, then (i,j,s_1,s_1,...,s_d,s_d) is 
         returned.
         """
@@ -340,7 +340,7 @@ class MultilayerNetwork(object):
             return len(list(self._iter_neighbors_out(node,dims)))
 
     def _get_degree_total_dir(self,node,dims=None):
-        """Returns the total degree of a _directed_ multilayer network.
+        """Return the total degree of a _directed_ multilayer network.
         """
         assert self.directed
         if dims==None:
@@ -349,7 +349,7 @@ class MultilayerNetwork(object):
             return len(list(self._iter_neighbors_total(node,dims)))
 
     def _get_degree_in_dir(self,node,dims=None):
-        """Returns the in-degree of a _directed_ multilayer network.
+        """Return the in-degree of a _directed_ multilayer network.
         """
         assert self.directed
         if dims==None:
@@ -388,7 +388,7 @@ class MultilayerNetwork(object):
         return sum(map(lambda n:self._get_link(self._nodes_to_link(node,n)),self._iter_neighbors_out(node,dims)))
 
     def _get_strength_total_dir(self,node, dims=None):
-        """Private method returning nodes total strenght (sum of in- and out-strength)."""
+        """Private method returning nodes total strength (sum of in- and out-strength)."""
         return self._get_strength_in(node,dims)+self._get_strength_out(node,dims)
 
 
@@ -549,7 +549,7 @@ class MultilayerNetwork(object):
 
 
     def get_layers(self,aspect=1):
-        """Returns the set of (elementary) layers (in a given aspect).
+        """Return the set of (elementary) layers (in a given aspect).
         """
         return self.slices[aspect]
 
@@ -567,13 +567,13 @@ class MultilayerNetwork(object):
                     yield node
 
     def __iter__(self):
-        """Iterates over all nodes.
+        """Iterate over all nodes.
         """
         for node in self.slices[0]:
             yield node
 
     def iter_node_layers(self):
-        """ Iterate over all node-layer pairs.
+        """Iterate over all node-layer pairs.
         """
         if self.fullyInterconnected:
             for nl in itertools.product(*map(lambda i:self.slices[i],range(len(self.slices)))):
@@ -590,7 +590,7 @@ class MultilayerNetwork(object):
 
 
     def iter_layers(self,aspect=None):
-        """ Iterate over all layers.
+        """Iterate over all layers.
 
[…]
