### 2024-07-01 7e6f64e
fix more MultilayerNode doc issues #26
