+++ pymnet/__init__.py
@@ -1,4 +1,4 @@
-from .net import MultilayerNetwork,MultiplexNetwork,MultilayerNode
+from .net import MultilayerNetwork,MultiplexNetwork
 from .models import er,conf,single_layer_er,single_layer_conf,er_partially_interconnected,full,full_multilayer,er_multilayer
 from .transforms import aggregate,subnet,supra_adjacency_matrix
 from .netio import read_ucinet
