+++ doc/reference.rst
@@ -18,6 +18,11 @@ Data Structures
 
   MultilayerNetwork
   MultiplexNetwork
+
+.. automodule:: pymnet.net
+.. autosummary::
+  :toctree: autogen
+
   MultilayerNode
 
 Network Models
