### 2024-07-01 mnets/pymnet#24 @dataspider
Thank you for these suggestions! We followed common hyphenation guidelines (e.g., [here](https://www.merriam-webster.com/grammar/hyphen-rules-open-closed-compound-words)) but will still change the title to your second suggestion, which nicely reflects the name of the library.
