### 2024-07-01 bafa8de
document net.MultilayerNode #26
