+++ pymnet/net.py
@@ -750,11 +750,6 @@ class MultilayerNode(object):
             layers=self.layers
         return self.mnet._get_strength_out(self.node,layers)
 
-    def layers(self,*layers):
-        """
-        """
-        return MultilayerNode(self.node,self.mnet,layers=layers)
-
     def iter_total(self):
         """Iterate over all neighbors.
         """
