### 2024-07-01 872e248
rename str -> strength #26
