+++ pymnet/tests/net_test.py
@@ -32,9 +32,9 @@ class TestNet(unittest.TestCase):
         self.assertEqual(net[1].deg(),1)
         self.assertEqual(net[2].deg(),2)
         self.assertEqual(net[3].deg(),1)
-        self.assertEqual(net[1].str(),1)
-        self.assertEqual(net[2].str(),2)
-        self.assertEqual(net[3].str(),1)
+        self.assertEqual(net[1].strength(), 1)
+        self.assertEqual(net[2].strength(), 2)
+        self.assertEqual(net[3].strength(), 1)
         self.assertEqual(list(net[4]),[])
 
         #Test alternative notations
@@ -59,8 +59,8 @@ class TestNet(unittest.TestCase):
         self.assertEqual(set(net[3]),set([1,2]))
         self.assertEqual(net[1].deg(),1)
         self.assertEqual(net[2].deg(),1)
-        self.assertEqual(net[1].str(),1)
-        self.assertEqual(net[2].str(),1)
+        self.assertEqual(net[1].strength(), 1)
+        self.assertEqual(net[2].strength(), 1)
         self.assertEqual(list(net[4]),[])
 
         self.assertEqual(net[1][2],0)
@@ -76,24 +76,24 @@ class TestNet(unittest.TestCase):
         self.assertEqual(set(net[3].iter_out()),set([1,2]))
         self.assertEqual(net[1].deg_out(),1)
         self.assertEqual(net[2].deg_out(),1)
-        self.assertEqual(net[1].str_out(),1)
-        self.assertEqual(net[2].str_out(),1)
+        self.assertEqual(net[1].strength_out(), 1)
+        self.assertEqual(net[2].strength_out(), 1)
         self.assertEqual(list(net[4].iter_out()),[])
 
         self.assertEqual(list(net[1].iter_in()),[3])
         self.assertEqual(set(net[3].iter_in()),set([1,2]))
         self.assertEqual(net[1].deg_in(),1)
         self.assertEqual(net[2].deg_in(),1)
-        self.assertEqual(net[1].str_in(),1)
-        self.assertEqual(net[2].str_in(),1)
+        self.assertEqual(net[1].strength_in(), 1)
+        self.assertEqual(net[2].strength_in(), 1)
         self.assertEqual(list(net[4].iter_in()),[])
 
         self.assertEqual(list(net[1].iter_total()),[3])
         self.assertEqual(set(net[3].iter_total()),set([1,2]))
         self.assertEqual(net[1].deg_total(),1)
         self.assertEqual(net[2].deg_total(),1)
-        self.assertEqual(net[1].str_total(),1)
-        self.assertEqual(net[2].str_total(),1)
+        self.assertEqual(net[1].strength_total(), 1)
+        self.assertEqual(net[2].strength_total(), 1)
         self.assertEqual(list(net[4].iter_total()),[])
         
 
@@ -117,9 +117,9 @@ class TestNet(unittest.TestCase):
         self.assertEqual(net[1].deg(),1)
         self.assertEqual(net[2].deg(),2)
         self.assertEqual(net[3].deg(),1)
-        self.assertEqual(net[1].str(),2)
-        self.assertEqual(net[2].str(),3)
-        self.assertEqual(net[3].str(),1)
+        self.assertEqual(net[1].strength(), 2)
+        self.assertEqual(net[2].strength(), 3)
+        self.assertEqual(net[3].strength(), 1)
         self.assertEqual(list(net[4]),[])
 
         self.assertEqual(list(net[1].iter_total()),[2])
[…]
