### 2024-07-01 mnets/pymnet#26 @dataspider
Docs are online [here](https://mnets.github.io/pymnet/autogen/pymnet.net.MultilayerNode.html#pymnet.net.MultilayerNode) and `str` has been renamed to `strength`.
