### 2024-07-01 #6930 @pitmonticone
I have a question for @danielskatz.

In one of the final sections of the paper, the authors cite several scientific publications (e.g., kivela2014, cozzo2015structure, kivela2017) as examples of projects that have adopted the open-source software package ⁠ pymnet ⁠. However, the first commit in the ⁠ pymnet ⁠ GitHub repository is dated October 2020. This suggests that the software was not publicly available as an open-source package, or even as supplementary code, at the time those earlier publications were written.

How does JOSS handle citations of projects that purportedly used code before it was publicly available as an open-source package?

Specifically:
- ⁠Should the authors provide additional context or evidence to clarify how the software was accessed and used in these earlier works?
- ⁠⁠Is there a requirement for the authors to disclose whether an earlier, non-public version of the software was used by these cited projects?

Your guidance on addressing this issue would be greatly appreciated, as I am not at all sure how to evaluate this.

Thank you in advance for your assistance.
