### 2024-07-01 #6930 @danielskatz
@pitmonticone - I don't see an issue here. The paper is about the software, and projects that have used it are just statement about those projects seeing some value in it. The license and status of the software at that time probably doesn't have anything to do with the value projects saw in using it. You can ask the authors about this, if you want, but I personally don't think the how they accessed it is necessary to add in the paper. JOSS certainly doesn't have any policies on this.
