### 2024-07-02 45a34d3
remove city names from paper
