### 2024-07-02 #6930 @ercco
@danielskatz According to this reviewer comment: https://github.com/mnets/pymnet/issues/24, we have updated the paper title to one suggested by the reviewer (`pymnet`: A Python Library for Multilayer Networks). However, the titles of the submission and this review issue, etc. are unchanged. I didn't find any method of editing the title post-submission on the JOSS submission page. I suppose we finish the review under the old title, and for the actual publication, the title can be updated to match the new paper title?
