### 2024-07-02 mnets/pymnet#30 @nwlandry
You might try adding a Github action to test the notebooks (For example, see our [Github action](https://github.com/xgi-org/xgi/blob/main/.github/workflows/test.yml) in XGI)
