### 2024-07-02 mnets/pymnet#34 @nwlandry
[REVIEW]: Use black and isort to format
There are several long lines ([this one](https://github.com/mnets/pymnet/blob/433f1dbb10d614defa3f6ac0777c2ba8390c8071/pymnet/net.py#L182), for example) which should be wrapped using a formatter such as black.

In addition, using isort would help standardize import statement order.

openjournals/joss-reviews#6930
