### 2024-07-02 mnets/pymnet#36 @nwlandry
[REVIEW] Add to multilayer introduction and make it more visible.
I really liked the [overview](https://mnets.github.io/pymnet/overview.html) page, but it would be great to 1) link to it on the README 2) add a visual(s) of multilayer networks for newcomers and 3) clarify the language surrounding "aspects". I understand that (like in [Multilayer Networks](https://doi.org/10.1093/comnet/cnu016)), this is pretty standard terminology, but it might be more consistent to rename `aspects` to `num_layers` or something similar.

openjournals/joss-reviews#6930
