### 2024-07-02 mnets/pymnet#37 @nwlandry
[REVIEW] Improve contribution guide
I see that there is a brief contribution guide in the README. It would be helpful to add a few issues with the "Good first issues" (For example, see [NetworkX](https://github.com/networkx/networkx/issues?q=is%3Aopen+is%3Aissue+label%3A%22Good+First+Issue%22), and link to this label in the contribution section.

In the future, it would be great to make the contribution guide a bit more step-by-step (e.g., fork the repository, make your changes, check test coverage, format code, add docs (is there a docstring format?), make a PR, etc.)

openjournals/joss-reviews#6930
