### 2024-07-02 mnets/pymnet#38 @nwlandry
[REVIEW] Add an example of an empirical dataset
The tutorials are really helpful, but it would be great if there was a tutorial on working with an empirical dataset (See [Netzschleuder](https://networks.skewed.de/?search=Multilayer), for example).

openjournals/joss-reviews#6930
