### 2024-07-02 mnets/pymnet#39 @nwlandry
[REVIEW] Remove setup.py
Is there a reason for having both pyproject.toml and setup.py? Can you remove setup.py?

https://github.com/openjournals/joss-reviews/issues/6930
