### 2024-07-02 #6930 @danielskatz
I've changed the title of this issue.  And the title in the generated PDF seems mostly ok now, in terms of the words. However, please remove the encoding of the software's name, i.e., remove the backquotes around pymnet in the title in the .md file. These will not be handled correctly in the processing for the final DOI and metadata that will be submitted to Crossref.
