+++ pymnet/diagnostics.py
@@ -1,8 +1,9 @@
-from .net import MultilayerNetwork,MultiplexNetwork
-from .transforms import subnet,aggregate,threshold
-import heapq,itertools
+from .net import MultilayerNetwork, MultiplexNetwork
+from .transforms import subnet, aggregate, threshold
+import heapq, itertools
 
-def degs(net,degstype="distribution"):
+
+def degs(net, degstype="distribution"):
     """Returns the degree distribution of a multilayer network.
 
     If the network has more than 1 aspect the degree distribution is returned for
@@ -20,21 +21,22 @@ def degs(net,degstype="distribution"):
        corresponding values are degrees of those nodes.
 
     """
-    if net.aspects==0:
-        the_iterator=net
+    if net.aspects == 0:
+        the_iterator = net
     else:
-        the_iterator=net.iter_node_layers()
-    degs={}
-    if degstype=="distribution":
+        the_iterator = net.iter_node_layers()
+    degs = {}
+    if degstype == "distribution":
         for node in the_iterator:
-            d=net[node].deg()
-            degs[d]=degs.get(d,0)+1
-    elif degstype=="nodes":
+            d = net[node].deg()
+            degs[d] = degs.get(d, 0) + 1
+    elif degstype == "nodes":
         for node in the_iterator:
-            degs[node]=net[node].deg()
+            degs[node] = net[node].deg()
     else:
         raise Exception("Invalid degstype parameter.")
     return degs
+
 
 def density(net):
     """Returns the density of the network.
@@ -43,33 +45,33 @@ def density(net):
     of possible edges in a general multilayer network with the same set of nodes and
     layers.
     """
-    if len(net)==0:
+    if len(net) == 0:
         return 0
 
-    if net.fullyInterconnected:        
-        nl=len(net.get_layers(0))
+    if net.fullyInterconnected:
+        nl = len(net.get_layers(0))
         for a in range(net.aspects):
-            nl=nl*len(net.get_layers(a+1))
+            nl = nl * len(net.get_layers(a + 1))
         if net.directed:
-            pedges=nl*(nl-1)
+            pedges = nl * (nl - 1)
         else:
-            pedges=(nl*(nl-1))/2
-            
-    return len(net.edges)/float(pedges)
+            pedges = (nl * (nl - 1)) / 2
+
+    return len(net.edges) / float(pedges)
 
 
 def multiplex_density(net):
-    """Returns a dictionary of densities of each intra-layer network of a multiplex network.
-    """
-    assert isinstance(net,MultiplexNetwork)
-    d={}
+    """Returns a dictionary of densities of each intra-layer network of a multiplex network."""
+    assert isinstance(net, MultiplexNetwork)
+    d = {}
     for layer in net.iter_layers():
-        d[layer]=density(net.A[layer])
+        d[layer] = density(net.A[layer])
     return d
 
-def multiplex_degs(net,degstype="distribution"):
+
+def multiplex_degs(net, degstype="distribution"):
     """Returns a dictionary of degree distributions of each intra-layer network of a multiplex network.
-    
+
     Parameters
     ----------
     net : MultiplexNetwork
@@ -82,15 +84,16 @@ def multiplex_degs(net,degstype="distribution"):
        corresponding values are degrees of those nodes.
 
     """
-    assert isinstance(net,MultiplexNetwork)
-    
-    d={}
+    assert isinstance(net, MultiplexNetwork)
+
+    d = {}
     for layer in net.iter_layers():
-        d[layer]=degs(net.A[layer],degstype=degstype)
+        d[layer] = degs(net.A[layer], degstype=degstype)
     return d
 
+
 def overlap_degs(net):
[…]
