+++ pymnet/tests/isomorphisms_test.py
@@ -6,10 +6,9 @@ from pymnet import net, transforms, models, isomorphisms
 
 
 def random_relabel(net, relabel_aspects):
-    """Randomly relabel some aspects of the input network.
-    """
+    """Randomly relabel some aspects of the input network."""
     labelings = []
-    for a in range(net.aspects+1):
+    for a in range(net.aspects + 1):
         if a in relabel_aspects:
             elayers = list(net.slices[a])
             random.shuffle(elayers)
@@ -19,8 +18,7 @@ def random_relabel(net, relabel_aspects):
             labelings.append({})
 
     if net.aspects >= 1:
-        return transforms.relabel(net, nodeNames=labelings[0],
-                                  layerNames=labelings[1:])
+        return transforms.relabel(net, nodeNames=labelings[0], layerNames=labelings[1:])
     else:
         return transforms.relabel(net, nodeNames=labelings[0])
 
@@ -39,13 +37,16 @@ class TestIsomorphisms(unittest.TestCase):
         results = {}
         if backend in isomorphisms.comparison_backends:
             results["comparison"] = isomorphisms.is_isomorphic(
-                net1, net2, allowed_aspects=allowed_aspects, backend=backend)
+                net1, net2, allowed_aspects=allowed_aspects, backend=backend
+            )
         if backend in isomorphisms.complete_invariant_backends:
             c1 = isomorphisms.get_complete_invariant(
-                net1, allowed_aspects=allowed_aspects, backend=backend)
+                net1, allowed_aspects=allowed_aspects, backend=backend
+            )
             c2 = isomorphisms.get_complete_invariant(
-                net2, allowed_aspects=allowed_aspects, backend=backend)
-            results["complete_invariant"] = (c1 == c2)
+                net2, allowed_aspects=allowed_aspects, backend=backend
+            )
+            results["complete_invariant"] = c1 == c2
 
         # Check that there was at least one way to get the result
         # This backend doesn't do anything...
@@ -53,7 +54,7 @@ class TestIsomorphisms(unittest.TestCase):
 
         # Check that the results are consistent
         # all_same= (len(set(results.values()))==1)
-        all_same = (len(set((results[k] for k in results))) == 1)
+        all_same = len(set((results[k] for k in results))) == 1
         self.assertTrue(all_same)
 
         # return results.values()[0]
@@ -98,55 +99,95 @@ class TestIsomorphisms(unittest.TestCase):
         isomorphisms.nxbackend.AuxiliaryGraphBuilderNX(netc, [0])
 
         # Network a Isomorphic to itself
-        self.assertTrue(self.is_isomorphic_multimethod(
-            neta, neta, allowed_aspects=[], backend=backend))
-        self.assertTrue(self.is_isomorphic_multimethod(
-            neta, neta, allowed_aspects=[0], backend=backend))
-        self.assertTrue(self.is_isomorphic_multimethod(
-            neta, neta, allowed_aspects=[1], backend=backend))
-        self.assertTrue(self.is_isomorphic_multimethod(
-            neta, neta, allowed_aspects=[0, 1], backend=backend))
+        self.assertTrue(
+            self.is_isomorphic_multimethod(
+                neta, neta, allowed_aspects=[], backend=backend
+            )
+        )
+        self.assertTrue(
+            self.is_isomorphic_multimethod(
+                neta, neta, allowed_aspects=[0], backend=backend
+            )
+        )
+        self.assertTrue(
+            self.is_isomorphic_multimethod(
+                neta, neta, allowed_aspects=[1], backend=backend
+            )
+        )
+        self.assertTrue(
+            self.is_isomorphic_multimethod(
+                neta, neta, allowed_aspects=[0, 1], backend=backend
+            )
+        )
 
         # Network a vertex-isomorphic to network b
[…]

[…]
