### 2024-07-02 f3ef36f
Sort imports with isort + add dev dependencies to toml #34
