+++ pymnet/tests/__init__.py
@@ -1,15 +1,16 @@
-from .net_test import test_net
 from .cc_test import test_cc
 from .diagnostics_test import test_diagnostics
 from .io_test import test_io
+from .isomorphisms_test import test_isomorphisms
 from .models_test import test_models
+from .net_test import test_net
+from .sampling_test import test_sampling
 from .transforms_test import test_transforms
 from .visuals_test import test_visuals
-from .isomorphisms_test import test_isomorphisms
-from .sampling_test import test_sampling
 
 try:
     import networkx
+
     from .nxwrap_test import test_nxwrap
 
     nximported = True
