+++ pymnet/sampling/__init__.py
@@ -1,7 +1,4 @@
 """Package for sampling multilayer subgraphs.
 """
 
-from . import esu
-from . import reqs
-from . import creators
-from . import dumb
+from . import creators, dumb, esu, reqs
