### 2024-07-02 mnets/pymnet#40 @nwlandry
[REVIEW] Condense code
Can you condense this [line](https://github.com/mnets/pymnet/blob/f3ef36f3f964defce55a3e081d45feab2cc96ce7/pymnet/net.py#L11) to
```python
try:
    from collections import MutableMapping
except ImportError:
    from collections.abc import MutableMapping
```

I don't love the `pass` statements. With the above, if neither work, it will throw an error.

[openjournals/joss-reviews/#6930](https://github.com/openjournals/joss-reviews/issues/6930)
