### 2024-07-02 mnets/pymnet#41 @nwlandry
[REVIEW] Mixing of snakecase and camelcase
This is certainly not essential to this review, but for greater consistency, I would choose snakecase or camelcase and stick with it. Right now, there is a mix of both.

https://github.com/openjournals/joss-reviews/issues/6930
