### 2024-07-02 mnets/pymnet#42 @nwlandry
[REVIEW] Standardized docstrings
Not necessary for this review, but in the future, it could be useful to standardize the docstrings. One example of a standard is [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html).

https://github.com/openjournals/joss-reviews/issues/6930
