### 2024-07-02 mnets/pymnet#43 @nwlandry
[REVIEW] Edits to JOSS paper
Overall the paper looks good. Some comments:
* The "Statement of Need" section was clearly written and helpful.
* There might not be space, but it would be cool to have a visualization of a small empirical multilayer network.
* In the summary, there are a few long and confusing sentences. For example, in the sentence "Multilayer networks (Kivelä et al., 2014) generalize graphs to capture the rich network data often associated with complex systems, allowing us to study a broad range of phenomena using a common representation, using the same multilayer tools and methods" can you remove the comma or end with "using the same representations, tools, and methods." In addition, can you rewrite "For many systems, there is a need to represent richer network data, such as multidimensional features, several types of interactions, numerous layers of hierarchy, and multiple modalities, to capture the structure and dynamics of the systems more accurately than simple graphs would allow." to something like "Many systems contain nodes with multidimensional features, interactions of different types, different levels of hierarchy, and multiple modalities, which are more faithfully captured with multilayer networks than simple graphs."
* The "Main Features and Examples" section was well-written and helpful.

https://github.com/openjournals/joss-reviews/issues/6930
