### 2024-07-03 c76d7f1
Resolves #40
