### 2024-07-03 e3a8aa8
Remove python2 remnants in "Using NetworkX functions" tutorial and add note on random seed #30
