### 2024-07-03 240f289
Tutorial edits wip #30
