+++ doc/tutorial/installing.rst
@@ -1,7 +1,7 @@
 Downloading and installing
 ==========================
 
-The easiest way to install pymnet is via pip:
+The easiest way to install :code:`pymnet` is via pip:
 
 .. code-block:: bash
 
@@ -13,7 +13,12 @@ Afterwards, the installation should work out-of-the-box.
 Alternatively, you can download the latest source files directly from our `GitHub Repository <https://github.com/mnets/pymnet>`_ as a zip file,
 or you can use git to clone the repository, for example, via HTTPS:
 
-.. code-block:: none
+.. code-block:: bash
 
-   $ git clone https://github.com/mnets/pymnet.git
+    $ git clone https://github.com/mnets/pymnet.git
 
+...or via SSH using a password-protected SSH key:
+
+.. code-block:: bash
+
+    $ git@github.com:mnets/pymnet.git
