### 2024-07-03 mnets/pymnet#30 @dataspider
RST Tutorials reviewed for python2 remnants + other update needs:
- [x] index
- [x] installing
- [ ] isomorphisms
- [ ] networktypes
- [x] nx
- [ ] visualizing
