### 2024-07-03 c82ccfa
Tutorial edits wip #30
