+++ doc/tutorial/isomorphisms.rst
@@ -6,37 +6,41 @@ Isomorphisms and automorphisms
 Requirements
 ------------
 
-The Multilayer Networks Library uses third party software to solve the graph-isomorphism problems. That is, some of the functionality of **pymnet** depends on having these software installed. Some of the functionality is redundant in a way that optional software packages can be used to solve the same problem, and only one of them is needed to be installed to gain access to that functionality.
+:code:`pymnet` uses third-party software to solve graph-isomorphism problems, i.e., some of its functionality depends on having this software installed. Some of the functionality is redundant in a way that optional software packages can be used to solve the same problem, and only one of them is needed to be installed to gain access to that functionality.
 
-- NetworkX : Available through *pip*. Allows the usage of following functions: *is_isomorphic*, *get_isomorphism*.
-- `bliss-blind` : Installed as a dependency on MacOS, X64 Windows and X64 Linux machines, has to be manually compiled and installed on other system using the command `python -m pip install bliss-bind`. Allows the usage of following functions: *is_isomorphic*, *get_isomorphism*, *get_automorphism_generators*, *get_complete_invariant*.
+- :code:`networkx` : Available through :code:`pip`. Allows the usage of the following functions: :code:`is_isomorphic`, :code:`get_isomorphism`.
+- :code:`bliss-blind` : Installed as a dependency on MacOS, X64 Windows and X64 Linux machines, has to be manually compiled and installed on other system using the command :code:`python -m pip install bliss-bind`. Allows the usage of the following functions: :code:`is_isomorphic`, :code:`get_isomorphism`, :code:`get_automorphism_generators`, :code:`get_complete_invariant`.
 
 Setup
 -----
 
-The Multilayer Networks Library can deal with generalized graph isomorphisms introduced in this `article <http://arxiv.org/abs/1309.7233>`_. This tutorial introduces the basic multilayer network isomorphism functionality in this library.
+:code:`pymnet` can deal with generalized graph isomorphisms introduced in `this article <https://doi.org/10.1093/comnet/cnu016>`_. This tutorial introduces the basic multilayer-network isomorphism functionality available in :code:`pymnet`.
 
-Lets start by creating a multilayer social network:
+Let's start by creating a multilayer social network:
 
->>> import pymnet
->>> net_social = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
->>> net_social["Alice", "Bob", "Friends"] = 1
->>> net_social["Alice", "Carol", "Friends"] = 1
->>> net_social["Bob", "Carol", "Friends"] = 1
->>> net_social["Alice", "Bob", "Married"] = 1
->>> fig_social = pymnet.draw(net_social, show=False, layout="spectral", layerPadding=0.2)
+.. code-block:: python
+
+    >>> import pymnet
+    >>> net_social = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
+    >>> net_social["Alice", "Bob", "Friends"] = 1
+    >>> net_social["Alice", "Carol", "Friends"] = 1
+    >>> net_social["Bob", "Carol", "Friends"] = 1
+    >>> net_social["Alice", "Bob", "Married"] = 1
+    >>> fig_social = pymnet.draw(net_social, show=False, layout="spectral", layerPadding=0.2)
 
 .. figure::  small_mplex_notaligned_social.png
    :align:   center
 
 We want to compare networks, so lets create another network. A transportation network this time:
 
[…]
