### 2024-07-03 mnets/pymnet#28 @cloner174
Hi!
The error message typically happens when the default backend does not support interactive plotting. But you have already installed PyQt6, so it seems the backend is not set correctly.
To resolve this, I suggest following steps, respectively!
1. Ensure that the backend is available in your environment:
```python
import matplotlib
print(matplotlib.rcsetup.all_backends)
```
2. Ensure PyQt6 is Installed Correctly:
```python
from PyQt6.QtWidgets import QApplication, QLabel
app = QApplication([])
label = QLabel("Congratulation!!!!!It's Okkkk")
label.show()
app.exec()
```
3. Ensure that Matplotlib is set to correct backend:
```python
import matplotlib
matplotlib.use('QtAgg') #or'Qt5Agg'

from pymnet import *
net = models.er_multilayer(5, 2, 0.2)
fig = draw(net, show=True)
```
4. Use Environment Variables:
```bash
export MPLBACKEND=QtAgg  #Unix-based
set MPLBACKEND=QtAgg  #Windows or'Qt5Agg'
```
5. Update All Packages:
```bash
conda update --all
```
6. Test in a Different Environment:
```bash
conda create -n test_ python=3.12 matplotlib pyqt6 pymnet
conda activate test_
```
