### 2024-07-03 mnets/pymnet#29 @cloner174
Hi!
Additionally:
0. verify that it is installed correctly:
```bash
pip list | grep bliss-bind
```
1. check for conflicting packages:
```bash
conda list
```
2. use a new environment.
