### 2024-07-03 3535425
remove duplicated sections from networktypes.rst
