### 2024-07-03 bac41b4
fix windows x64 platform relax matplotlib version
