+++ pyproject.toml
@@ -28,11 +28,11 @@ classifiers = [
 ]
 
 dependencies = [
-  "matplotlib >= 3.8.0",
+  "matplotlib >= 3.7.0",
   "numpy >= 1.26.1",
   "scipy >= 1.11.1",
   "networkx >= 2.0, < 4.0",
-  "bliss-bind >= 0.2.0 ; (sys_platform == 'darwin' or ( (platform_machine == 'x86_64' and sys_platform == 'linux') or (platform_machine == 'amd64' and sys_platform == 'win32') )) and python_version >= '3.8'"
+  "bliss-bind >= 0.2.0 ; (sys_platform == 'darwin' or ( (platform_machine == 'x86_64' and sys_platform == 'linux') or (platform_machine == 'AMD64' and sys_platform == 'win32') )) and python_version >= '3.8'"
 ]
 
 [project.urls]
