### 2024-07-03 9f280f6
relax numpy and scipy dependencies
