### 2024-07-03 mnets/pymnet#29 @arashbm
Fix pushed in bac41b4 and will be in the next release. Seems to be correctly downloading `bliss-bind` as a dependency on x64 Windows GitHub runners.

@ClaudMor Until the next release you can manually install `bliss-bind` (`pip install -U bliss-bind`)  as a workaround.
