### 2024-07-03 17ced70
remove setup.py (Fix #39)
