+++ setup.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python
-# vim: set fileencoding=utf-8
-
-from setuptools import setup
-
-
-setup()
