### 2024-07-03 f61b24e
remove remenants of bitbucket era
