+++ tox.ini
@@ -1,12 +1,0 @@
-[tox]
-envlist=py27-nx111,py38-{nx111,nx25}
-skip_missing_interpreters = true
-[testenv]
-deps=
-  nx111: networkx==2.0
-  nx25: networkx==2.5
-  matplotlib
-  numpy
-  scipy
-commands=
-	python -m pymnet.tests
