### 2024-07-03 fe438c0
add HACKING.md
