+++ .github/workflows/tests.yml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   build:
-    name: run tests on ${{ matrix.os }}
+    name: run tests on ${{ matrix.os }} + Python ${{ matrix.python }}
     runs-on: ${{ matrix.os }}
 
     strategy:
