### 2024-07-03 mnets/pymnet#37 @arashbm
@dataspider I added a simple [HACKING.md](https://github.com/mnets/pymnet/blob/master/HACKING.md) with some of the essentials but will have to add some more information, e.g. formatting commands, code coverage etc.
