### 2024-07-03 3f72406
remove tox MANIFEST file
