+++ MANIFEST.in
@@ -1,2 +1,0 @@
-recursive-include pymnet *.py
-recursive-include pymnet/tests *.py
