### 2024-07-03 7451b2c
minor edits on models
