+++ pymnet/models.py
@@ -1,12 +1,6 @@
-"""Functions for generating multilayer and multiplex networks using various network models.
+"""Functions for generating multilayer and multiplex networks using various
+network models.
 """
-
-##### Compatibility for Python 2/3
-try:
-    xrange
-except NameError:
-    xrange = range
-######
 
 import math
 import random
@@ -24,21 +18,23 @@ def single_layer_conf(net, degs, degstype="distribution"):
     degs : dict
        Degrees of the network. See degstype parameter.
     degstype : string
-       If 'distribution', then degs parameter gives the degree distribution. I.e.,
-       keys are degrees, and corresponding values are number of nodes with the given degree.
-       If 'nodes', then degs parameter gives node degrees. I.e, keys are node names and
-       corresponding values are degrees of those nodes.
+       If 'distribution', then degs parameter gives the degree distribution,
+       i.e., keys are degrees, and corresponding values are number of nodes
+       with the given degree. If 'nodes', then degs parameter gives node
+       degrees, i.e, keys are node names and corresponding values are degrees
+       of those nodes.
 
     Notes
     -----
     The algorithm used here is similar to the one in article:
-    B.D McKay, N.C Wormald 'Uniform Generation of Random Regular Graphs of Moderate Degree'
-    Journal of Algorithms 11, pages 52-67 (1990)
+    B.D McKay, N.C Wormald 'Uniform Generation of Random Regular Graphs of
+    Moderate Degree' Journal of Algorithms 11, pages 52-67 (1990)
 
-    The difference between the algorithm presented in the article and the one in this
-    function is that the random restarts are not implemented here. This means that the
-    sampled networks are not exactly statistically uniform. However, if the degrees
-    are small compared to the number of nodes the error is likely to be small.
+    The difference between the algorithm presented in the article and the one
+    in this function is that the random restarts are not implemented here. This
+    means that the sampled networks are not exactly statistically uniform.
+    However, if the degrees are small compared to the number of nodes the error
+    is likely to be small.
     """
     stubs = []
     selfedges = {}
@@ -83,7 +79,8 @@ def single_layer_conf(net, degs, degstype="distribution"):
     for s in range(int(len(stubs) / 2)):
         node1, node2 = sorted([stubs[2 * s], stubs[2 * s + 1]])
 
-        edgetoindex[(node1, node2)] = edgetoindex.get((node1, node2), []) + [2 * s]
+        edgetoindex[(node1, node2)] = edgetoindex.get((node1, node2), []) + \
+            [2 * s]
 
         if net[node1, node2] != 0:
             multiedges.add((node1, node2))
@@ -99,13 +96,16 @@ def single_layer_conf(net, degs, degstype="distribution"):
             while repeat:
                 # select two edges at random
                 e1i, e2i = map(
-                    lambda x: 2 * x, random.sample(xrange(int(len(stubs) / 2)), 2)
+                    lambda x: 2 * x,
+                    random.sample(range(int(len(stubs) / 2)), 2)
                 )
-                c = [node, stubs[e1i], stubs[e1i + 1], stubs[e2i], stubs[e2i + 1]]
+                c = [node, stubs[e1i], stubs[e1i + 1],
+                     stubs[e2i], stubs[e2i + 1]]
                 n2, n3 = sorted([c[1], c[2]])
                 n4, n5 = sorted([c[3], c[4]])
                 if len(set(c)) == len(c):
-                    if (n2, n3) not in multiedges and (n4, n5) not in multiedges:
+                    if (n2, n3) not in multiedges and \
+                            (n4, n5) not in multiedges:
                         if (
[…]
