### 2024-07-03 7894f1c
raw docstring for latex
