### 2024-07-03 39ab4c3
minor style edits in netio + write_json test
