### 2024-07-04 74d1fd5
remove the old PyBliss backend
It is not compatible with Python 3, and we now have a minimum supported version
of 3.8+.
