### 2024-07-04 6684b11
remove tests related to NetworkX 1.0
We no longer support NetworkX 1.0
