### 2024-07-04 0f5e23f
Rename basic guide by @arashbm to CONTRIBUTING.md for better integration with GitHub features #37
