+++ pymnet/__init__.py
@@ -1,15 +1,38 @@
 from . import graphlets, isomorphisms
-from .cc import (avg_lcc_aw, cc_barrat, cc_barrett, cc_onnela, cc_sequence,
-                 cc_zhang, elementary_cycles, gcc_aw, gcc_zhang, lcc, lcc_aw,
-                 lcc_brodka, sncc_aw)
+from .cc import (
+    avg_lcc_aw,
+    cc_barrat,
+    cc_barrett,
+    cc_onnela,
+    cc_sequence,
+    cc_zhang,
+    elementary_cycles,
+    gcc_aw,
+    gcc_zhang,
+    lcc,
+    lcc_aw,
+    lcc_brodka,
+    sncc_aw,
+)
 from .diagnostics import degs, density, multiplex_degs, multiplex_density
-from .isomorphisms import (get_automorphism_generators, get_complete_invariant,
-                           get_isomorphism, is_isomorphic)
-from .models import (conf, er, er_multilayer, er_partially_interconnected,
-                     full, full_multilayer, single_layer_conf, single_layer_er)
+from .isomorphisms import (
+    get_automorphism_generators,
+    get_complete_invariant,
+    get_isomorphism,
+    is_isomorphic,
+)
+from .models import (
+    conf,
+    er,
+    er_multilayer,
+    er_partially_interconnected,
+    full,
+    full_multilayer,
+    single_layer_conf,
+    single_layer_er,
+)
 from .net import MultilayerNetwork, MultiplexNetwork
-from .netio import (read_edge_files, read_ucinet, write_edge_files, write_json,
-                    write_ucinet)
+from .netio import read_edge_file, read_ucinet, write_edge_files, write_json
 from .transforms import aggregate, subnet, supra_adjacency_matrix
 from .visuals import draw, webplot
 
