### 2024-07-04 27f3040
RST tutorials visualizing, networktypes #30
