+++ doc/tutorial/networktypes.rst
@@ -1,139 +1,179 @@
 Network types
 =============
 
-
 Monoplex networks
 -----------------
 
 Start by importing the library:
 
->>> from pymnet import *
+.. code-block:: python
 
-MultilayerNetwork is the basic network class in the library -- all other types of networks are special case of it. In order to get a monoplex network object
-you can simply construct a MultilayerNetwork with 0 aspects.
+    >>> from pymnet import *
 
->>> net = MultilayerNetwork(aspects=0)
+:code:`MultilayerNetwork` is the basic network class in the library – all other types of networks are special cases of it. In order to get a monoplex network object, you can simply construct a :code:`MultilayerNetwork` with 0 aspects.
 
-You can now start adding nodes to the network with the 'add_node' method:
+.. code-block:: python
 
->>> net.add_node(1)
->>> net.add_node(2)
+    >>> net = MultilayerNetwork(aspects=0)
 
-Iterating the network object will yield all the nodes.
+You can now start adding nodes to the network with the :code:`add_node` method:
 
->>> list(net)
-[1, 2]
+.. code-block:: python
 
-You can access node objects of the network with the get syntax. The node objects can be used for example to get the degree or strength of the node.
+    >>> net.add_node(1)
+    >>> net.add_node(2)
 
->>> net[1].deg()
-0
+Iterating over the network object will yield all the nodes.
 
-You can now start adding edges to your network. There are two basic ways of doing this. First is the syntax which resembles setting elements in an adjacency matrix:
+.. code-block:: python
 
->>> net[1,2] = 1
+    >>> list(net)
+    [1, 2]
 
-In the second way, you first access a node object and then get the edge pointing to a neighbor of that node:
+You can access the node objects of the network with the :code:`get` syntax. The node objects can be used, for example, to get the degree or strength of the node.
 
->>> net[1][2] = 1
+.. code-block:: python
 
-These to ways are equivalent, but the difference will become important when we move on to networks which are not monoplex ones. You can also add edges between nodes that you have not explicitly created, and in this case the missing nodes are created automatically:
+    >>> net[1].deg()
+    0
 
->>> net[1,3] = 1
->>> list(net)
-[1, 2, 3]
+You can now start adding edges to your network. There are two basic ways of doing this. As the first option, you can use syntax which resembles setting elements in an adjacency matrix:
 
-You can use similar syntax to check if the edges exist and access their weights. The default weight value for non-existing edge is 0.
+.. code-block:: python
 
->>> net[1,2]
-1
->>> net[1][3]
-1
->>> net[2,3]
-0
+    >>> net[1, 2] = 1
 
-The network is undirected by default which means that adding an edge in one direction adds it automatically to both directions.
+As a second option, you can first access a node object and then get the edge pointing to a neighbor of that node:
 
->>> net[2,1]
-1
+.. code-block:: python
+
+    >>> net[1][2] = 1
+
+These to ways are equivalent, but the difference will become important when we move on to networks that are not monoplex anymore. You can also add edges between nodes that you have not explicitly created, and in this case, the missing nodes are created automatically:
+
+.. code-block:: python
+
+    >>> net[1, 3] = 1
+    >>> list(net)
+    [1, 2, 3]
+
+You can use similar syntax to check if the edges exist and to access their weights. The default weight value for non-existing edges is 0.
+
+.. code-block:: python
+
+    >>> net[1, 2]
+    1
+    >>> net[1][3]
+    1
[…]
