+++ doc/tutorial/visualizing.rst
@@ -3,67 +3,80 @@
 Visualizing networks
 ====================
 
-The Multilayer Network Library can produce high quality images of multilayer and multiplex networks. The main method of producing network images in the library uses Matplotlib as a backend. These images can be saved both in vector formats (such as pdf or svg) and raster formats (such as png). Alternatively, one can produce pictures of networks using a method that uses Javascript and D3 as a backend for displaying figures. These interactive figures can be viewed with any modern browser. (Also, NetworkX is used for some layouts).
+:code:`pymnet` can produce high-quality images of multilayer and multiplex networks. The main method for producing network images in the library uses :code:`matplotlib`  as a backend. These images can be saved in both vector formats (such as pdf or svg) and raster formats (such as png). Alternatively, one can produce pictures of networks using a method that uses Javascript and D3 as a backend for displaying figures. These interactive figures can be viewed with any modern browser. (Also, :code:`networkx` is used for some layouts).
 
-Drawing networks with the library is easy and is usually done with the "draw" method. Consider, for example, the following code:
+Drawing networks with the library is easy and is usually done with the :code:`draw` method. Consider, for example, the following code:
 
->>> from pymnet import *
->>> net = models.er_multilayer(5,2,0.2)
->>> fig = draw(net)
+.. code-block:: python
 
-The first line of this code imports the multilayer network. The second line creates a random node-aligned multilayer network with 5 nodes and 2 layers with each node-layer tuple connected to each other with a probability of 0.2. The third line then creates a picture of that network. 
+    >>> from pymnet import *
+    >>> net = models.er_multilayer(5, 2, 0.2)
+    >>> fig = draw(net)
 
-Note that running this code doesn't actually show the figure, but it's stored in the computers memory at this point. You can save the figure to a file, for example with the following command:
+The first line of this code imports :code:`pymnet`. The second line creates a random node-aligned multilayer network with 5 nodes and 2 layers, where two node-layers are connected to each other with a probability of 0.2. The third line then creates a picture of that network.
 
->>> fig.savefig("net.pdf")
+Note that running this code doesn't actually show the figure, but it's stored in the computers memory at this point. You can save the figure to a file, for example, with the following command:
 
-Alternatively you can view the network straight away by telling the draw method that you want the figure to be shown:
+.. code-block:: python
 
->>> fig = draw(net,show=True)
+    >>> fig.savefig("net.pdf")
 
-The produced figure looks like this:
+Alternatively, you can view the network straight away by telling the draw method that you want the figure to be shown:
+
+.. code-block:: python
+
+    >>> fig = draw(net, show=True)
+
+The figure produced looks like this:
 
 .. figure::  ermlayer1.png
    :align:   center
 
-Multiplex network figures can also be produced. For example running the following code...
+Multiplex-network figures can also be produced. For example, running the following code...
 
->>> fig=draw(er(10,3*[0.4]),layout="spring")
+.. code-block:: python
 
-produces a following picture of multiplex ER network:
+    >>> fig = draw(er(10, 3*[0.4]), layout="spring")
+
+produces the following picture of a multiplex Erdos-Renyi network:
 
 .. figure:: ermplex1.png
    :align:   center
 
-There are multiple ways of customizing the figures. For documentation look at the reference for the draw method. Here is an example usage of the draw methods that uses several of the customization options:
[…]
