### 2024-07-05 5413f12
add test case for fully connected ER single layer
