### 2024-07-05 2684a44
use ValueError for parameter related exceptions
