### 2024-07-06 54579bb
add tests for optimised barrat cc
