+++ pymnet/tests/cc_test.py
@@ -23,6 +23,7 @@ class TestCC(unittest.TestCase):
         self.assertEqual(cc.cc_zhang(n, 1), 1.0)
         self.assertEqual(cc.cc_onnela(n, 1), 1.0)
         self.assertEqual(cc.cc_barrat(n, 1), 1.0)
+        self.assertEqual(cc.cc_barrat_optimized(n, 1), 1.0)
 
     def test_unweighted_flat_simple(self):
         n = net.MultilayerNetwork(aspects=0)
@@ -35,6 +36,7 @@ class TestCC(unittest.TestCase):
         self.assertEqual(cc.cc_zhang(n, 1), 1.0 / 3.0)
         self.assertEqual(cc.cc_onnela(n, 1), 1.0 / 3.0)
         self.assertEqual(cc.cc_barrat(n, 1), 1.0 / 3.0)
+        self.assertEqual(cc.cc_barrat_optimized(n, 1), 1.0 / 3.0)
 
     def test_weighted_flat_simple(self):
         pass  # TODO
@@ -56,6 +58,7 @@ class TestCC(unittest.TestCase):
 
         # barrett
         self.assertEqual(cc.cc_barrett(n, 1, an), 128.0 / 96.0)
+        self.assertEqual(cc.cc_barrett_optimized(n, 1, an), 128.0 / 96.0)
         self.assertEqual(cc.cc_barrett(n, 1, an), cc.cc_barrett_explicit(n, 1))
 
         # cc seq
@@ -598,7 +601,8 @@ class TestCC(unittest.TestCase):
                     cc.cc_cycle_vector_bf(net, snode, layer),
                 )
 
-        # Test that Barrett clustering coeffiecient is consistent across different versions
+        # Test that Barrett clustering coeffiecient is consistent across
+        # different versions
         anet = pymnet.transforms.aggregate(net, 1)
         for snode in net.slices[0]:
             self.assertEqual(
@@ -607,6 +611,10 @@ class TestCC(unittest.TestCase):
             self.assertEqual(
                 cc.cc_barrett_optimized(net, snode, anet),
                 cc.cc_barrett(net, snode, anet),
+            )
+            self.assertEqual(
+                cc.cc_barrett(net, snode, anet),
+                cc.cc_barrett_optimized(net, snode, anet)
             )
 
     def test_unweighted_nonglobalnodes_consistency_er(self):
