### 2024-07-06 a90b4b6
fix typo in cc_barrette_optimized tests
