### 2024-07-06 afcedb4
simple edgefile io test
