### 2024-07-06 9f27f30
fix write_edge_files test
