### 2024-07-06 7a4da26
Refactor, restrict, and test read_edge_list
