### 2024-07-06 3aca1e1
Add missing test to suite
