### 2024-07-06 fcf34c8
write_edge_file function to match read_edge_file format
