### 2024-07-06 4793d5a
complete docs of current io functions; closes #33
