### 2024-07-06 c76ae06
improve contribution guide; closes #37
