### 2024-07-06 03e4f6c
use name mangling to hide mnet in MultilayerNode; raise TypeError when trying to set non-numeric values to edges; closes #27
