### 2024-07-06 f719e5f
replace `Exception`s with proper builtin subclass
