### 2024-07-06 mnets/pymnet#29 @ClaudMor
I confirm that the error above no longer happens after your fix.

@arashbm if it still useful, here's what your command returns:

```nothing
PS C:\Users\claud> python -c "import sys, platform; print(sys.platform, platform.machine())"
win32 AMD64
```

I couldn't find the `bliss_bind` package on PyPi because on the documentation it is listed as `bliss_blind` ([link](https://mnets.github.io/pymnet/tutorial/isomorphisms.html)).
