### 2024-07-06 mnets/pymnet#29 @arashbm
Thanks for verifying. This also confirms that we fixed the correct issue

> I couldn't find the bliss_bind package on PyPi because on the documentation it is listed as bliss_blind (link).

Understandable. As far as Python is concerned,  both options should point to the same thing. You should be able to pip install the underscore one and it should still work. [PyPI normalizes names to the hyphen version](https://packaging.python.org/en/latest/specifications/name-normalization/#name-normalization) and that's what is used in the package name and anywhere we refer to the package.

But since you can't easily import a module with a hyphen in it, in code we used the underscore version to refer to the module offered by that package.

At least that has been the goal. I'll check to make sure if it is consistent.
