### 2024-07-06 mnets/pymnet#30 @ClaudMor
Thanks @dataspider for translating [this](https://mnets.github.io/pymnet/tutorial/nx.html) tutorial to python 3.

Anyway, I suggest changing the following line:

```python
print(map(nx.number_connected_components,models.er(1000,3*[0.005]).A.values()))
```

to

```python
print(map(nx.number_connected_components,pymnet.models.er(1000,3*[0.005]).A.values()))
```

to avoid an Exception being thrown. Moreover, its output is still:

```nothing
<map object at 0x0000013B1166CEB0>
```

instead of the documented

```nothing
[10, 9, 5]
```
