### 2024-07-06 27d2eb7
fix networkx tutorial
