+++ doc/tutorial/nx.rst
@@ -5,7 +5,13 @@ Using NetworkX functions
 
 Start by importing the library:
 
->>> from pymnet import nx
+>>> from pymnet import nx, models
+
+
+For the sake of reproducability let's explicitly seed the ranom number generator here:
+
+>>> import random
+>>> random.seed(42)
 
 You can then run any NetworkX function from the pymnet.nx module. For example, you can produce the Karate Club network with the following command.
 
@@ -13,13 +19,12 @@ You can then run any NetworkX function from the pymnet.nx module. For example, y
 
 This will produce a native Pymnet multilayer network object with 0 aspects (i.e., a monoplex network). To confirm this, try:
 
->>> print(net)
+>>> net
 <pymnet.net.MultilayerNetwork at 0x7f5b550eaa10>
+>>> net.aspects
+0
 
 You can also pass Pymnet objects as arguments to NetworkX functions in a similar way. This is handy, for example, when analyzing monoplex structures of intra-layer networks of multiplex networks. For example, producing a multiplex network with three Erdos-Renyi intra-layer networks using Pymnet and calculating the number of connected components in each layer can be done with the following command:
 
->>> print(map(nx.number_connected_components,models.er(1000,3*[0.005]).A.values()))
-[10, 9, 5]
-
-This result was generated with :code:`random.seed(42))` (after runnign :code:`import random`).
-Your output may differ depending on the seed of the random-number generator used by the :code:`random` module.
+>>> {name: nx.number_connected_components(layer) for name, layer in models.er(1000, 3*[0.005]).A.items()}
+{0: 10, 1: 9, 2: 5}
