### 2024-07-06 3d4b1e4
move around text in the nxwrap tutorial slightly
