+++ doc/tutorial/nx.rst
@@ -7,12 +7,6 @@ Start by importing the library:
 
 >>> from pymnet import nx, models
 
-
-For the sake of reproducability let's explicitly seed the ranom number generator here:
-
->>> import random
->>> random.seed(42)
-
 You can then run any NetworkX function from the pymnet.nx module. For example, you can produce the Karate Club network with the following command.
 
 >>> net = nx.karate_club_graph()
@@ -23,6 +17,11 @@ This will produce a native Pymnet multilayer network object with 0 aspects (i.e.
 <pymnet.net.MultilayerNetwork at 0x7f5b550eaa10>
 >>> net.aspects
 0
+
+For the sake of reproducability in the next example, let's explicitly seed the ranom number generator here:
+
+>>> import random
+>>> random.seed(42)
 
 You can also pass Pymnet objects as arguments to NetworkX functions in a similar way. This is handy, for example, when analyzing monoplex structures of intra-layer networks of multiplex networks. For example, producing a multiplex network with three Erdos-Renyi intra-layer networks using Pymnet and calculating the number of connected components in each layer can be done with the following command:
 
