### 2024-07-06 mnets/pymnet#30 @arashbm
@ClaudMor Thanks, I manually fixed this tutorial for now, but tutorials will get a notebook revamp at some point soon.
