### 2024-07-06 mnets/pymnet#30 @dataspider
My bad (some of my local changes got lost in translation) – thanks @arashbm for fixing this (and yes, jupyter notebook versions are on the way)!
