### 2024-07-06 a498bd6
remove wildcard imports from old tutorials
