### 2024-07-06 b7abe63
add toml as docs requirement
