### 2024-07-09 #6930 @danielskatz
👋 @nwlandry - I see https://github.com/mnets/pymnet/issues/37 is closed. Does this allow you to check off [your last item](https://github.com/openjournals/joss-reviews/issues/6930#issuecomment-2192336197)?
