### 2024-07-10 #6930 @nwlandry
@danielskatz thanks for the heads up. I've checked off my last item, though I would recommend that the contributing guide be put on the top level instead of in the `.github` folder.
