### 2024-07-10 #6930 @ercco
Moving the contribution guide to the root of the repository? Yes, absolutely. We thank @nwlandry for all the excellent suggestions! The optional suggestions are also very good, so we have the goal of implementing them as well.
