### 2024-07-11 mnets/pymnet#44 @pitmonticone
[REVIEW]: Typos in `docs/overview`
Please consider fixing these typos in [`docs/overview`](https://mnets.github.io/pymnet/overview.html)

> Here we take **~e~** a similar approach

> We typically consider **~that~** the number of aspects, $d$, to be constant and not dependent on the size of the data.

> That is, the graph is a dictionary where each node [...] is a key ~and values~ whose value is another dictionary containing information about the neighbors of each node.

JOSS Review Issue: https://github.com/openjournals/joss-reviews/issues/6930
