### 2024-07-11 mnets/pymnet#45 @pitmonticone
[REVIEW]: Typos in `docs/tutorial/networktypes`
Please consider fixing these typos in [`docs/tutorial/networktypes`](https://mnets.github.io/pymnet/tutorial/networktypes.html):

> These **~to~ two** ways are equivalent, but [...]

> The multilayer networks can in theory be used to represent**~ed~** multiplex networks

JOSS Review Issue: https://github.com/openjournals/joss-reviews/issues/6930
