+++ doc/overview.rst
@@ -3,7 +3,7 @@
 Overview
 ========
 
-The data types in this library follow the mathematical definitions of various multilayer network types defined in the article `"Multilayer Networks", M. Kivela et al., arXiv:1309.7233 [physics.soc-ph] <http://arxiv.org/abs/1309.7233>`_. In the article a multilayer network is defined as a general mathematical structure and all the other types of networks are defined as special cases of that structure. Here we take e a similar approach and define a class MultilayerNetwork such that it represents the mathematical definition of the multilayer network. All the other network classes then inherit the MultilayerNetwork class. Currently we have the MultiplexNetwork class which represent multiplex networks as defined in the article. In the article there were several constraints defined for multiplex networks. Some of these constraints, such as "categorical" and "ordinal" couplings, are also implemented in this library. Instances of MultiplexNetwork that are constrained in this way can be implemented efficiently and the algorithms dealing with general multilayer networks can take advantage of the information that the network object is constrained. 
+The data types in this library follow the mathematical definitions of various multilayer network types defined in the article `"Multilayer Networks", M. Kivela et al., arXiv:1309.7233 [physics.soc-ph] <http://arxiv.org/abs/1309.7233>`_. In the article a multilayer network is defined as a general mathematical structure and all the other types of networks are defined as special cases of that structure. Here we take e a similar approach and define a class :code:`MultilayerNetwork` such that it represents the mathematical definition of the multilayer network. All the other network classes then inherit the :code:`MultilayerNetwork` class. Currently we have the :code:`MultiplexNetwork` class which represent multiplex networks as defined in the article. In the article there were several constraints defined for multiplex networks. Some of these constraints, such as "categorical" and "ordinal" couplings, are also implemented in this library. Instances of :code:`MultiplexNetwork` that are constrained in this way can be implemented efficiently and the algorithms dealing with general multilayer networks can take advantage of the information that the network object is constrained. 
 
 
 Computational efficiency
