### 2024-07-11 f402cf6
Update networktypes.rst
