+++ doc/tutorial/networktypes.rst
@@ -184,7 +184,7 @@ Multiplex networks
 
 The multilayer networks can in theory be used to represented multiplex networks, but in practice, it is often better to use a specialized class :code:`MultiplexNetwork` when dealing with multiplex networks.
 There are several reasons for this. First, the :code:`MultiplexNetwork` class offers an additional convenient interface for handling intra-layer networks.
-Second, the MultiplexNetwork class can take coupling rules as an input when it is constructed and use them to implicitly create the inter-layer edges when they are needed. This saves some memory and makes it easier to create networks with such coupling structures.
+Second, the :code:`MultiplexNetwork` class can take coupling rules as an input when it is constructed and use them to implicitly create the inter-layer edges when they are needed. This saves some memory and makes it easier to create networks with such coupling structures.
 Third, using a :code:`MultiplexNetwork` will let the functions in the library know that your multilayer network is a multiplex network. Some of the functions only work for multiplex networks, but even the ones that work for general multilayer networks can use the information to speed up the processing.
 
 The simplest multiplex network is the one with no coupling edges. You would create such an object with the following command:
