### 2024-07-11 738bc2a
Update nx.rst
