### 2024-07-13 7944f4d
Update networktypes.rst (fix #45)
Thanks to @pitmonticone for spotting the typos
