### 2024-07-14 8587bc0
first jupyter notebook tutorial with attempt at smooth unittest integration
