### 2024-07-14 d3abc46
move tutorial tests out of main test suite
