+++ pymnet/test_tutorials/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from . import test_all
+
+sys.exit(not test_all())
