+++ .github/workflows/tests.yml
@@ -35,7 +35,7 @@ jobs:
         run: python -m pip install -U pip
 
       - name: Install the package
-        run: python -m pip install .[test,tutorial]
+        run: python -m pip install .[test]
 
       - name: Run tests
         run: coverage run -m pymnet.tests
