+++ pymnet/test_tutorials/__init__.py
@@ -1,6 +1,0 @@
-from .tutorial_test import test_tutorials
-
-def test_all():
-    codes = []
-    codes.append(test_tutorials())
-    return all(codes)
