+++ pymnet/test_tutorials/__main__.py
@@ -1,5 +1,0 @@
-import sys
-
-from . import test_all
-
-sys.exit(not test_all())
