### 2024-07-14 ae78584
keep notebook tests local for now
