+++ pymnet/tests/tutorial_test.py
@@ -1,14 +1,15 @@
 import nbformat
 import unittest
 import subprocess
+import sys
 
 
 class TestTutorials(unittest.TestCase):
     def setUp(self):
-        self.networktypes = "tutorials/01_networktypes"
+        self.networktypes = "../tutorials/01_networktypes"
 
     def test_networktypes(self):
-        res = subprocess.call(
+        subprocess.call(
             f"jupyter nbconvert --to notebook --execute {self.networktypes}.ipynb",
             shell=True,
         )
@@ -21,9 +22,15 @@ class TestTutorials(unittest.TestCase):
         self.assertListEqual(out, out_original)
 
     def tearDown(self):
-        subprocess.call(f"rm tutorials/{self.networktypes}.nbconvert.ipynb", shell=True)
+        subprocess.call(
+            f"rm ../tutorials/{self.networktypes}.nbconvert.ipynb", shell=True
+        )
 
 
 def test_tutorials():
     suite = unittest.TestSuite()
     suite.addTest(TestTutorials("test_networktypes"))
+
+
+if __name__ == "__main__":
+    sys.exit(not test_tutorials())
