### 2024-07-14 7fc437b
move and rename notebooks for inclusion in docs via myst
