### 2024-07-14 f599f13
add myst_nb to requirements and conf (+ some code formatting)
