### 2024-07-14 293f284
add myst_nb to requirements
