+++ doc/requirements.txt
@@ -2,3 +2,4 @@ numpydoc == 1.6.0
 furo
 sphinx-copybutton
 toml
+myst_nb
