### 2024-07-14 10fbb4a
new docs setup
