+++ doc/tutorial_old/index.rst
@@ -1,4 +1,4 @@
-.. _tutorial:
+.. _tutorial_old:
 
 Tutorials
 =========
@@ -8,7 +8,6 @@ The following tutorials will help you with to get started with using the library
 .. toctree::
    :maxdepth: 5
 
-   installing
    networktypes
    visualizing
    isomorphisms
