### 2024-07-14 2840ad1
small adjustment of local test
