### 2024-07-15 254002d
remove old tutorial sources
