+++ doc/tutorial_old/index.rst
@@ -1,14 +1,0 @@
-.. _tutorial_old:
-
-Tutorials
-=========
-
-The following tutorials will help you with to get started with using the library.
-
-.. toctree::
-   :maxdepth: 5
-
-   networktypes
-   visualizing
-   isomorphisms
-   nx
