+++ doc/tutorial_old/networktypes.rst
@@ -1,267 +1,0 @@
-Network types
-=============
-
-Monoplex networks
------------------
-
-Start by importing the library:
-
-.. code-block:: python
-
-    >>> import pymnet
-
-:code:`MultilayerNetwork` is the basic network class in the library – all other types of networks are special cases of it. In order to get a monoplex network object, you can simply construct a :code:`MultilayerNetwork` with 0 aspects.
-
-.. code-block:: python
-
-    >>> net = pymnet.MultilayerNetwork(aspects=0)
-
-You can now start adding nodes to the network with the :code:`add_node` method:
-
-.. code-block:: python
-
-    >>> net.add_node(1)
-    >>> net.add_node(2)
-
-Iterating over the network object will yield all the nodes.
-
-.. code-block:: python
-
-    >>> list(net)
-    [1, 2]
-
-You can access the node objects of the network with the :code:`get` syntax. The node objects can be used, for example, to get the degree or strength of the node.
-
-.. code-block:: python
-
-    >>> net[1].deg()
-    0
-
-You can now start adding edges to your network. There are two basic ways of doing this. As the first option, you can use syntax which resembles setting elements in an adjacency matrix:
-
-.. code-block:: python
-
-    >>> net[1, 2] = 1
-
-As a second option, you can first access a node object and then get the edge pointing to a neighbor of that node:
-
-.. code-block:: python
-
-    >>> net[1][2] = 1
-
-These two ways are equivalent, but the difference will become important when we move on to networks that are not monoplex anymore. You can also add edges between nodes that you have not explicitly created, and in this case, the missing nodes are created automatically:
-
-.. code-block:: python
-
-    >>> net[1, 3] = 1
-    >>> list(net)
-    [1, 2, 3]
-
-You can use similar syntax to check if the edges exist and to access their weights. The default weight value for non-existing edges is 0.
-
-.. code-block:: python
-
-    >>> net[1, 2]
-    1
-    >>> net[1][3]
-    1
-    >>> net[2,3]
-    0
-
-The network is undirected by default, which means that adding an edge in one direction automatically adds it in both directions.
-
-.. code-block:: python
-
-    >>> net[2, 1]
-    1
-
-The node objects can be iterated, which will yield the list of neighbors of the node:
-
-.. code-block:: python
-
-    >>> list(net[1])
-    [2, 3]
-
-The edges can be removed by simply setting their value to 0:
-
-.. code-block:: python
-
-    >>> net[1, 3] = 0
-    >>> list(net[1])
-    [2]
-
-The edges of the networks can also be weighted with arbitrary numbers (except with the one corresponding to a missing edge). In weighted networks,
-the degree and the weighted degree, i.e. strength, of a node are different:
-
-.. code-block:: python
-
-    >>> net[1,3] = 2
-    >>> net[1].deg()
-    2
-    >>> net[1].strength()
-    3
-
-By default, all network objects are undirected. Directed network objects can be created by setting the keyword parameter :code:`directed` to :code:`True` in the constructor of the network object:
-
-.. code-block:: python
-
-    >>> dirnet = pymnet.MultilayerNetwork(aspects=0, directed=True)
-    >>> dirnet[1, 2] = 1
-    >>> dirnet[1, 2]
-    1
-    >>> dirnet[2, 1]
-    0
-    >>> net[2, 1]
-    1
-
-
-Multilayer networks
--------------------
-
-We are now ready to move to more general multilayer networks with an  arbitrary number of aspects. For simplicity, we will start with a network that has a single aspect.
-
-.. code-block:: python
-
[…]
