+++ doc/tutorial_old/isomorphisms.rst
@@ -1,122 +1,0 @@
-.. _isomorphisms_tutorial:
-
-Isomorphisms and automorphisms
-==============================
-
-Requirements
-------------
-
-:code:`pymnet` uses third-party software to solve graph-isomorphism problems, i.e., some of its functionality depends on having this software installed. Some of the functionality is redundant in a way that optional software packages can be used to solve the same problem, and only one of them is needed to be installed to gain access to that functionality.
-
-- :code:`networkx` : Available through :code:`pip`. Allows the usage of the following functions: :code:`is_isomorphic`, :code:`get_isomorphism`.
-- :code:`bliss-blind` : Installed as a dependency on MacOS, X64 Windows and X64 Linux machines, has to be manually compiled and installed on other system using the command :code:`python -m pip install bliss-bind`. Allows the usage of the following functions: :code:`is_isomorphic`, :code:`get_isomorphism`, :code:`get_automorphism_generators`, :code:`get_complete_invariant`.
-
-Setup
------
-
-:code:`pymnet` can deal with generalized graph isomorphisms introduced in `this article <https://doi.org/10.1093/comnet/cnu016>`_. This tutorial introduces the basic multilayer-network isomorphism functionality available in :code:`pymnet`.
-
-Let's start by creating a multilayer social network:
-
-.. code-block:: python
-
-    >>> import pymnet
-    >>> net_social = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
-    >>> net_social["Alice", "Bob", "Friends"] = 1
-    >>> net_social["Alice", "Carol", "Friends"] = 1
-    >>> net_social["Bob", "Carol", "Friends"] = 1
-    >>> net_social["Alice", "Bob", "Married"] = 1
-    >>> fig_social = pymnet.draw(net_social, show=False, layout="spectral", layerPadding=0.2)
-
-.. figure::  small_mplex_notaligned_social.png
-   :align:   center
-
-We want to compare networks, so lets create another network. A transportation network this time:
-
-.. code-block:: python
-
-    >>> net_transport = pymnet.MultiplexNetwork(couplings='categorical', fullyInterconnected=False)
-    >>> net_transport["Helsinki", "Turku", "Train"] = 1
-    >>> net_transport["Helsinki", "Tampere", "Train"] = 1
-    >>> net_transport["Turku", "Tampere", "Train"] = 1
-    >>> net_transport["Helsinki", "Turku", "Ferry"] = 1
-    >>> fig_transport = pymnet.draw(net_transport, show=False, layout="circular", layerPadding=0.2)
-
-.. figure::  small_mplex_notaligned_transport.png
-   :align:   center
-
-
-Isomorphisms
-------------
-
-Clearly the two above networks have the same structure but different labels of nodes and layers. In terms of multilayer network isomorphisms, this means that the two networks are node-layer isomorphic (but not node isomorphic or layer isomorphic). This can be confirmed using the *is_isomorphic* function as follows: 
-
-.. code-block:: python
-
-    >>> print(pymnet.is_isomorphic(net_social, net_transport, allowed_aspects=[0])) #node isomorphism
-    False
-
-    >>> print(pymnet.is_isomorphic(net_social, net_transport, allowed_aspects=[1])) #layer isomorphism
-    False
-
-    >>> print(pymnet.is_isomorphic(net_social, net_transport, allowed_aspects=[0, 1])) #node-layer isomorphism
-    True
-
[…]
