+++ doc/conf.py
@@ -36,6 +36,8 @@ extensions = [
     "myst_nb",
 ]
 
+nb_execution_mode = "force"
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
