+++ doc/conf.py
@@ -36,7 +36,7 @@ extensions = [
     "myst_nb",
 ]
 
-nb_execution_mode = "force"
+nb_execution_mode = "off"
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
