### 2024-07-15 mnets/pymnet#30 @dataspider
Tutorials are now migrated to Python 3 and translated (where applicable) into Jupyter Notebooks, which are included in the docs via MyST_NB; see, e.g., [here](https://mnets.github.io/pymnet/tutorials/isomorphisms.html).

@nwlandry Thank you for the pointer to your GitHub action for testing notebooks – I am hoping to do cell-by-cell output comparisons and integrate these into our suite of unit tests (stub [here](https://github.com/mnets/pymnet/blob/master/pymnet/tests/tutorial_test.py)) but have yet to set make this work in our GitHub-action test environment. But since the docs now show the results of true notebook executions, I'll treat this as a separate issue and close this one.
