### 2024-07-15 28efc75
attempt to fix interactive backend detection on windows #28
