### 2024-07-15 mnets/pymnet#28 @dataspider
@ClaudMor
Our current hunch is that the previous display-detection code did not work in (some or all) Windows environments.
It is no surprise that the code suggested above by someone else does not fix the problem because it imports pymnet after setting the matplotlib backend, and then pymnet will have just looked at os.environ.get("DISPLAY", "") again and set "Agg" as a backend although you have interactive backends available on your system.

Could you kindly try two things:
1. Running in a jupyter notebook (e.g., using one of the now-notebook tutorials)
Expected behavior: No problems (the display should be set automatically by jupyter)
2. Running again in your previous setup after reinstalling pymnet to include commit 28efc75 (linked to this issue)
Expected behavior: No problems (the commit should fix the display-detection problem)

I don't have access to Windows machines, so I'm stabbing a bit in the dark...
