### 2024-07-15 e6723d1
catch tornado-related runtime errors in updated backend-detection code under linux #28
