### 2024-07-15 mnets/pymnet#30 @nwlandry
@dataspider That sounds like a plan and certainly not essential for the review. Just a suggestion for future development.
