### 2024-07-15 cc63c06
more informative readme #36
