### 2024-07-15 bf48252
small readme fixes, more detail + image in overview #36
