+++ doc/overview.rst
@@ -3,39 +3,64 @@
 Overview
 ========
 
-The data types in this library follow the mathematical definitions of various multilayer network types defined in the article `"Multilayer Networks", M. Kivela et al., arXiv:1309.7233 [physics.soc-ph] <http://arxiv.org/abs/1309.7233>`_. In the article a multilayer network is defined as a general mathematical structure and all the other types of networks are defined as special cases of that structure. Here we take e a similar approach and define a class :code:`MultilayerNetwork` such that it represents the mathematical definition of the multilayer network. All the other network classes then inherit the :code:`MultilayerNetwork` class. Currently we have the :code:`MultiplexNetwork` class which represent multiplex networks as defined in the article. In the article there were several constraints defined for multiplex networks. Some of these constraints, such as "categorical" and "ordinal" couplings, are also implemented in this library. Instances of :code:`MultiplexNetwork` that are constrained in this way can be implemented efficiently and the algorithms dealing with general multilayer networks can take advantage of the information that the network object is constrained. 
+Multilayer networks as defined by `Kivelä et al. (2014) <https://doi.org/10.1093/comnet/cnu016>`_ generalize graphs to capture the rich network data often associated with complex systems, allowing us to study a broad range of phenomena using a common representation, using the same multilayer tools and methods.
+Formally, a *multilayer network* :math:`M` is defined as a quadruple :math:`M = (V_M, E_M, V,` **L**:math:`)`, where the sequence **L** :math:`= (L_a)_{𝑎=1}^d` defines sets :math:`L_a` of *elementary layers*, the set :math:`V` defines the *nodes* of the network, the *node-layers* are :math:`V_M ⊆ V × L_1 × ... × L_d`, and the *edges* :math:`E_M ⊆ V_M × V_M` are defined between node-layers.
+Put simply, a node-layer is an association of a node :math:`v ∈ V` with a layer :math:`∈ L ×...× L` of dimensionality :math:`d`, nodes can exist on an arbitrary number of layers, and edges can connect node-layers within layers and across arbitrary pairs of layers, which can differ in an arbitrary number of dimensions. The dimensions :math:`1, 2, ..., d` are called the *aspects* of the network (e.g., a two-aspect transport network could have *time* as its first aspect and *transport mode* as its second aspect).
+
+The data types in :code:`pymnet` mirror the formalization stated above. In `Kivelä et al. (2014) <https://doi.org/10.1093/comnet/cnu016>`_, a multilayer network is defined as a general mathematical structure, and all the other types of networks are defined as special cases of that structure.
+Here, we take a similar approach and define a class :code:`MultilayerNetwork` such that it represents the mathematical definition of the multilayer network. All the other network classes then inherit the :code:`MultilayerNetwork` class. Currently, we have the :code:`MultiplexNetwork` class, which represents multiplex networks as defined in the article. In the article, there were several constraints defined for multiplex networks. Some of these constraints, such as "categorical" and "ordinal" couplings, are also implemented in this library. Instances of :code:`MultiplexNetwork` that are constrained in this way can be implemented efficiently, and the algorithms dealing with general multilayer networks can take advantage of the information that the network object is constrained.
+
+For example, the following code constructs and visualizes a small multiplex social network:
+
+.. code-block:: python
+
+   from pymnet import *
+
+   net_social = MultiplexNetwork(couplings="categorical", fullyInterconnected=False)
+   net_social["Alice", "Bob", "Friends"] = 1
+   net_social["Alice", "Carol", "Friends"] = 1
+   net_social["Bob", "Carol", "Friends"] = 1
+   net_social["Alice", "Bob", "Married"] = 1
+
[…]
