### 2024-07-15 df66322
image display fix
