### 2024-07-15 2b71074
better figure embedding
