### 2024-07-15 ee96867
move misplaced figure
