### 2024-07-15 mnets/pymnet#36 @dataspider
Overview is now linked to the README, both README and overview contain network visuals, and the overview has been augmented to include the formal multilayer-network definition and an example of what "aspects" means.
Crucially, we cannot rename "aspects" to "number of layers" or the like because this would misrepresent the semantics and run counter to the established terminology (of which I personally am also not a fan, but oh well).
I hope this settles all sub-issues of this issue – but feel free to reopen in case you feel otherwise!
