### 2024-07-15 7ce19f1
include py3plex (closes #23)
