### 2024-07-15 7c338eb
add Aalto affiliation for cc
