### 2024-07-15 e0e32d9
trying to do away with backend detection on pymnet's side altogether, letting mpl do the job #28
