### 2024-07-15 19e2fd1
some unit tests for uncovered models
