### 2024-07-15 9a1b8de
some more unit tests for uncovered models
