+++ pymnet/models.py
@@ -79,8 +79,7 @@ def single_layer_conf(net, degs, degstype="distribution"):
     for s in range(int(len(stubs) / 2)):
         node1, node2 = sorted([stubs[2 * s], stubs[2 * s + 1]])
 
-        edgetoindex[(node1, node2)] = edgetoindex.get((node1, node2), []) + \
-            [2 * s]
+        edgetoindex[(node1, node2)] = edgetoindex.get((node1, node2), []) + [2 * s]
 
         if net[node1, node2] != 0:
             multiedges.add((node1, node2))
@@ -96,16 +95,13 @@ def single_layer_conf(net, degs, degstype="distribution"):
             while repeat:
                 # select two edges at random
                 e1i, e2i = map(
-                    lambda x: 2 * x,
-                    random.sample(range(int(len(stubs) / 2)), 2)
+                    lambda x: 2 * x, random.sample(range(int(len(stubs) / 2)), 2)
                 )
-                c = [node, stubs[e1i], stubs[e1i + 1],
-                     stubs[e2i], stubs[e2i + 1]]
+                c = [node, stubs[e1i], stubs[e1i + 1], stubs[e2i], stubs[e2i + 1]]
                 n2, n3 = sorted([c[1], c[2]])
                 n4, n5 = sorted([c[3], c[4]])
                 if len(set(c)) == len(c):
-                    if (n2, n3) not in multiedges and \
-                            (n4, n5) not in multiedges:
+                    if (n2, n3) not in multiedges and (n4, n5) not in multiedges:
                         if (
                             net[node, n2] == 0
                             and net[node, n4] == 0
@@ -127,16 +123,13 @@ def single_layer_conf(net, degs, degstype="distribution"):
             while repeat:
                 # select two edges at random
                 e1i, e2i = map(
-                    lambda x: 2 * x,
-                    random.sample(range(int(len(stubs) / 2)), 2)
+                    lambda x: 2 * x, random.sample(range(int(len(stubs) / 2)), 2)
                 )
-                c = [n1, n2, stubs[e1i], stubs[e1i + 1],
-                     stubs[e2i], stubs[e2i + 1]]
+                c = [n1, n2, stubs[e1i], stubs[e1i + 1], stubs[e2i], stubs[e2i + 1]]
                 n3, n4 = sorted([c[2], c[3]])
                 n5, n6 = sorted([c[4], c[5]])
                 if len(set(c)) == len(c):
-                    if (n3, n4) not in multiedges and \
-                            (n5, n6) not in multiedges:
+                    if (n3, n4) not in multiedges and (n5, n6) not in multiedges:
                         if (
                             net[n1, n3] == 0
                             and net[n2, n4] == 0
@@ -154,8 +147,7 @@ def single_layer_conf(net, degs, degstype="distribution"):
                             net[n1, n5] = 1
                             net[n2, n6] = 1
                             si1, si2 = sorted(
-                                [edgetoindex[n1, n2].pop(),
-                                 edgetoindex[n1, n2].pop()]
+                                [edgetoindex[n1, n2].pop(), edgetoindex[n1, n2].pop()]
                             )
                             stubs[si1], stubs[si1 + 1] = sorted([n1, n3])
                             stubs[si2], stubs[si2 + 1] = sorted(
@@ -292,11 +284,9 @@ def conf(degs, degstype="distribution", couplings=("categorical", 1.0)):
                     nodeAligned = False
                 nnodes = lnnodes
         else:
-            raise ValueError(
-                "degstype should be either `distribution' or `nodes'")
+            raise ValueError("degstype should be either `distribution' or `nodes'")
 
[…]
