+++ README.md
@@ -51,5 +51,5 @@ Before you get started, please check out our [contribution guide](CONTRIBUTING.m
 
 ## Asking Questions
 
-* For bugs, feature requests, etc., please open an issue in our [issue tracker](https://github.com/mnets/pymnet/issues).
+* For bugs, feature requests, etc., please use [GitHub issues][github-issues].
 * Otherwise, feel free to contact the main developer: [Mikko Kivelä](http://www.mkivela.com/)
