### 2024-07-16 381991f
typo in drawcore doc
