### 2024-07-16 d215691
real-world data tutorial preprocessing + docs #38
