### 2024-07-16 mnets/pymnet#38 @dataspider
A tutorial using real-world data is now available [here](https://mnets.github.io/pymnet/tutorials/realworld.html).

As the datasets tagged "multilayer" on Netzschleuder (and also those available in most other places) had neither cleanly documented provenance nor reliable (i.e., archival DOI, not personal homepage) upstream URLs, I ended up using an excerpt from the [CITES data](https://trade.cites.org/) on trade in endangered species compiled by UNEP, which is also redistributed in the tutorial folder along with the preprocessing code.

(I am hoping to make the full data available for multilayer-network analysis via Zenodo, too, such that it can readily be included also in Netzschleuder and other aggregators, but I want to check with CITES/UNEP first).
