### 2024-07-16 mnets/pymnet#38 @nwlandry
@dataspider --- this is perfect! Sorry that it was more involved to find such a dataset, but I think that it makes pymnet more compelling. Nice work.
