### 2024-07-16 mnets/pymnet#36 @nwlandry
@dataspider --- perfect, I agree that this is done.
