### 2024-07-16 a4b3d69
real-world example in paper wip #43
