### 2024-07-16 0e029cc
paper compilation things
