+++ paper/paper.md
@@ -91,9 +91,9 @@ In the multilayer-network model representing this data, our nodes are countries 
 More details on the data provenance and our preprocessing steps are available in the `pymnet` repository.
 
 In \autoref{fig:citestrade}, we visualize the data with `pymnet`, only drawing undirected edges representing aggregate trade volumes of at least one million specimens resp. one thousand kilograms, scaling nodes by degree, and mapping trade volume to edge widths and edge colors.
-We observe an expansion of trade from 2010 to 2020 in both trade volume and diversity of trade partners, which is particularly pronounced for trade in meat.
+Between 2010 and 2020, we observe roughly equal levels of trade in live specimens, but a pronounced increase in both trade volume and diversity of trade partners for trade in meat.
 
-![Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right). \label{fig:citestrade}](https://github.com/mnets/pymnet/blob/publication/paper/figs/cites-trades_commercial_live-meat_2010.png?raw=true "Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right)."){ width=75% } 
+![Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right). \label{fig:citestrade}](https://github.com/mnets/pymnet/blob/publication/paper/figs/cites-trades_commercial_live-meat.png?raw=true "Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right)."){ width=75% } 
 
 # Installation and Usage
 
