### 2024-07-16 mnets/pymnet#43 @dataspider
The real-world example is now also integrated in the main text, as shown in [this paper-compilation artifact](https://github.com/mnets/pymnet/actions/runs/9961635183/artifacts/1707501646).
With this, all your requested edits should be implemented (let me know if you think otherwise).
Thank you for your constructive suggestions!
