### 2024-07-17 5c68939
some code reformatting and adding basic black and isort config to pyproject.toml
