+++ pymnet/__init__.py
@@ -34,8 +34,8 @@ from .models import (
 from .net import MultilayerNetwork, MultiplexNetwork
 from .netio import (
     read_edge_file,
-    write_edge_file,
     read_ucinet,
+    write_edge_file,
     write_edge_files,
     write_json,
 )
