+++ pymnet/sampling/dumb.py
@@ -2,8 +2,11 @@
 
 import itertools
 
-from .reqs import (default_calculate_required_lengths, default_check_reqs,
-                   relaxed_check_reqs)
+from .reqs import (
+    default_calculate_required_lengths,
+    default_check_reqs,
+    relaxed_check_reqs,
+)
 
 
 def dumb_enumeration(
