+++ pymnet/graphlets/__init__.py
@@ -8,9 +8,14 @@ Sallamari Sallmen, Tarmo Nurmi, and Mikko Kivelä. "Graphlets in multilayer netw
 from . import graphlet_measures
 from . import graphlets as graphlets_module
 from . import independent_equations as independent_equations_module
-from .graphlet_measures import (GCD, GCM, GCD_matrix, orbit_counts,
-                                orbit_counts_all, orbit_numbers,
-                                ordered_orbit_list)
-from .graphlets import (automorphism_orbits, graphlets, list_orbits,
-                        orbit_equations)
+from .graphlet_measures import (
+    GCD,
+    GCM,
+    GCD_matrix,
+    orbit_counts,
+    orbit_counts_all,
+    orbit_numbers,
+    ordered_orbit_list,
+)
+from .graphlets import automorphism_orbits, graphlets, list_orbits, orbit_equations
 from .independent_equations import independent_equations, redundant_orbits
