### 2024-07-17 df14053
unify citation syntax for bliss papers
