### 2024-07-17 402b31c
use hash on complete invariants instead of str
