+++ pymnet/graphlets/graphlets.py
@@ -29,10 +29,10 @@ def graphlets(n, layers, n_l=None, couplings=None, allowed_aspects="all"):
     -------
     nets : dict (key: n_nodes, value: list of MultiplexNetwork objects)
         graphlets
-    invariants : dict (key: str(complete invariant), value: tuple(index in 'nets': n_nodes, index in the list of multiplex networks))
+    invariants : dict (key: complete invariant, value: tuple(index in 'nets': n_nodes, index in the list of multiplex networks))
         complete invariants of the graphlets, the values can be used to match the graphlets in 'nets'
     """
-    if n_l == None:
+    if n_l is None:
         n_l = len(layers)
 
     nets = {}
@@ -42,17 +42,18 @@ def graphlets(n, layers, n_l=None, couplings=None, allowed_aspects="all"):
     for net_layers in itertools.combinations(layers, n_l):
         layer_combs = layer_combinations(net_layers)
         for layer_comb in layer_combs:
-            net = pymnet.MultiplexNetwork(couplings=couplings, fullyInterconnected=True)
+            net = pymnet.MultiplexNetwork(
+                couplings=couplings, fullyInterconnected=True)
             for layer in layer_comb:
                 net[0, 1, layer] = 1
 
             for layer in net_layers:
                 net.add_layer(layer)
 
-            ci = pymnet.get_complete_invariant(net, allowed_aspects=allowed_aspects)
-            ci_s = str(ci)
-            if not ci_s in invariants:
-                invariants[ci_s] = (2, len(nets2))
+            ci = pymnet.get_complete_invariant(
+                net, allowed_aspects=allowed_aspects)
+            if ci not in invariants:
+                invariants[ci] = (2, len(nets2))
                 nets2.append(net)
 
     nets[2] = nets2
@@ -77,13 +78,13 @@ def graphlets(n, layers, n_l=None, couplings=None, allowed_aspects="all"):
 
                         for layer in net_layers:
                             new_net.add_layer(layer)
-                        # check if isomorphic with a previous graph & add only if not isomorphic
+                        # check if isomorphic with a previous graph & add only
+                        # if not isomorphic
                         ci = pymnet.get_complete_invariant(
                             new_net, allowed_aspects=allowed_aspects
                         )
-                        ci_s = str(ci)
-                        if not ci_s in invariants:
-                            invariants[ci_s] = (i + 1, len(nets_i_1))
+                        if ci not in invariants:
+                            invariants[ci] = (i + 1, len(nets_i_1))
                             nets_i_1.append(new_net)
 
         nets[i + 1] = nets_i_1
@@ -106,8 +107,9 @@ def automorphism_orbits(nets, allowed_aspects="all"):
     -------
     auts : dict (key: (n_nodes, net_index, node), value: node_orbit_index)
         Automorphism orbits. 'n_nodes' is the key in 'nets'. 'net_index' is the
-        index in the list of networks. 'node' is the node name in the given network.
-        'node_orbit_index' gets the same value for nodes that are in the same orbit.
+        index in the list of networks. 'node' is the node name in the given
+        network. 'node_orbit_index' gets the same value for nodes that are in
+        the same orbit.
     """
 
     auts = dd()
@@ -189,7 +191,7 @@ def orbit_equations(n, nets, auts, invs, allowed_aspects="all"):
         graphlets, as returned by graphlets
[…]
