+++ pyproject.toml
@@ -32,7 +32,7 @@ dependencies = [
   "numpy >= 1.23.0",
   "scipy >= 1.10.0",
   "networkx >= 2.0, < 4.0",
-  "bliss-bind >= 0.2.0 ; (sys_platform == 'darwin' or ( (platform_machine == 'x86_64' and sys_platform == 'linux') or (platform_machine == 'AMD64' and sys_platform == 'win32') )) and python_version >= '3.8'"
+  "bliss-bind >= 0.3.0 ; (sys_platform == 'darwin' or ( (platform_machine == 'x86_64' and sys_platform == 'linux') or (platform_machine == 'AMD64' and sys_platform == 'win32') )) and python_version >= '3.8'"
 ]
 
 [project.urls]

[…]
