### 2024-07-17 07caf2c
use absolute link for readme image
