### 2024-07-17 f52560a
bump version to 0.3.0
