### 2024-07-18 6b828b8
add initial graphlet tests
