### 2024-07-18 ed6c0b0
add tests for the rest of user-facing functions in graphlets file
