+++ pymnet/tests/graphlets_test.py
@@ -39,10 +39,129 @@ class TestGraphlets(unittest.TestCase):
         assert match_ref_1
         assert match_ref_2
 
+    def test_automorphism_orbits(self):
+        nets, invs = graphlets.graphlets(
+            n=3,
+            layers=["a", "b", "c"],
+            n_l=2,
+            couplings="categorical",
+            allowed_aspects="all",
+        )
+        auts = graphlets.automorphism_orbits(nets, allowed_aspects="all")
+        target_auts = {
+            (3, 0, 2): 1,
+            (3, 9, 0): 0,
+            (3, 4, 0): 0,
+            (3, 3, 1): 0,
+            (3, 4, 1): 1,
+            (3, 3, 0): 0,
+            (3, 4, 2): 1,
+            (3, 3, 2): 0,
+            (3, 8, 0): 0,
+            (3, 7, 1): 0,
+            (3, 8, 1): 1,
+            (3, 2, 2): 2,
+            (3, 7, 0): 0,
+            (2, 0, 1): 0,
+            (3, 8, 2): 1,
+            (3, 1, 2): 1,
+            (3, 2, 0): 0,
+            (3, 7, 2): 2,
+            (3, 1, 1): 1,
+            (3, 2, 1): 1,
+            (3, 1, 0): 0,
+            (3, 6, 2): 1,
+            (3, 5, 2): 1,
+            (3, 6, 0): 0,
+            (3, 5, 1): 1,
+            (3, 6, 1): 1,
+            (2, 1, 0): 0,
+            (3, 5, 0): 0,
+            (2, 1, 1): 0,
+            (2, 0, 0): 0,
+            (3, 0, 0): 0,
+            (3, 9, 2): 0,
+            (3, 0, 1): 1,
+            (3, 9, 1): 0,
+        }
+        assert auts == target_auts
+
+    def test_list_orbits(self):
+        nets, invs = graphlets.graphlets(
+            n=3,
+            layers=["a", "b", "c"],
+            n_l=2,
+            couplings="categorical",
+            allowed_aspects="all",
+        )
+        auts = graphlets.automorphism_orbits(nets, allowed_aspects="all")
+        orbit_lists = graphlets.list_orbits(auts)
+        target_orbit_lists = {
+            2: [(2, 0, 0), (2, 1, 0)],
+            3: [
+                (3, 0, 1),
+                (3, 9, 0),
+                (3, 4, 0),
+                (3, 3, 0),
+                (3, 4, 1),
+                (3, 8, 0),
+                (3, 7, 0),
+                (3, 8, 1),
+                (3, 2, 2),
+                (3, 1, 1),
+                (3, 2, 0),
+                (3, 7, 2),
+                (3, 2, 1),
+                (3, 1, 0),
+                (3, 6, 1),
+                (3, 5, 1),
+                (3, 6, 0),
+                (3, 5, 0),
+                (3, 0, 0),
+            ],
+        }
+        assert set(orbit_lists[2]) == set(target_orbit_lists[2])
+        assert set(orbit_lists[3]) == set(target_orbit_lists[3])
+
+    def test_orbit_equations(self):
+        nets, invs = graphlets.graphlets(
+            n=3,
+            layers=["a", "b", "c"],
+            n_l=2,
+            couplings="categorical",
+            allowed_aspects="all",
+        )
+        auts = graphlets.automorphism_orbits(nets, allowed_aspects="all")
+        orbit_eqs = graphlets.orbit_equations(
+            n=3, nets=nets, auts=auts, invs=invs, allowed_aspects="all"
+        )
+        target_orbit_eqs = {
[…]
