### 2024-07-18 3ea2989
add tests for independent_equations user-facing functions
