### 2024-07-19 73ffaaa
add tests for orbit_counts_all and orbit_counts
