### 2024-07-20 mnets/pymnet#28 @ClaudMor
Dear @dataspider,

I confirm that the jupyter notebooks properly work. Also the scripts work, provided I install pymnet including of [28efc75](https://github.com/mnets/pymnet/commit/28efc75aaca5f3f73c666035634e2468f08070ea) and I subsequently also install PyQt6.
