+++ paper/paper.md
@@ -90,7 +90,7 @@ To keep matters simple, we analyze an excerpt from the CITES database, investiga
 In the multilayer-network model representing this data, our nodes are countries or territories importing and exporting endangered species (represented by their ISO-2 codes), our aspects are years (two elementary layers: 2010 and 2020) and trade terms (two elementary layers: "live" and "meat"), and our edges represent bilateral trades, measured in the number of specimens ("live") or the number of kilograms ("meat") traded. 
 More details on the data provenance and our preprocessing steps are available in the `pymnet` repository.
 
-In \autoref{fig:citestrade}, we visualize the data with `pymnet`, only drawing undirected edges representing aggregate trade volumes of at least one million specimens resp. one thousand kilograms, scaling nodes by degree and highlighting OECD countries in red, as well as mapping trade volume to edge widths and edge colors.
+In \autoref{fig:citestrade}, we visualize the data with `pymnet`, only drawing undirected edges representing aggregate trade volumes of at least one million specimens ("live") or one thousand kilograms ("meat"), scaling nodes by degree and highlighting OECD countries in red, as well as mapping trade volume to edge widths and edge colors.
 Between 2010 and 2020, we observe roughly equal levels of trade in live specimens, but a pronounced increase in both trade volume and diversity of trade partners for trade in meat.
 
 ![Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right). \label{fig:citestrade}](https://github.com/mnets/pymnet/blob/publication/paper/figs/cites-trades_commercial_live-meat.png?raw=true "Trade in live specimens (top) and meat (bottom) of endangered species in the years 2010 (left) and 2020 (right)."){ width=95% } 
