### 2024-07-22 #6930 @danielskatz
👋 @pitmonticone - It looks like you just have two items left in [your checklist](https://github.com/openjournals/joss-reviews/issues/6930#issuecomment-2197211251). Can you say what's holding you back from checking these items?
