### 2024-07-22 #6930 @ClaudMor
> 👋 @ClaudMor - I see your checklist is complete. Can you confirm that you think this is ready to accept?

Hi @danielskatz ,

I am done with my review and I think this paper represents a valid contribution. I therefore encourage the authors to satisfy the remaining other reviewers' issues and recommend acceptance!

Nice work!
