### 2024-07-22 8f47824
minor changes in paper.md
