### 2024-07-23 da21a3d
minor changes in paper.bib
