+++ paper/paper.bib
@@ -44,36 +44,28 @@
       doi={10.48550/arXiv.2308.00083}
 }
 
-@InProceedings{junttila2011,
-  author =    {Tommi Junttila and Petteri Kaski},
-  title =     {Conflict Propagation and Component Recursion for Canonical Labeling},
-  booktitle = {Theory and Practice of Algorithms in (Computer) Systems -- First International {ICST} Conference, {TAPAS} 2011, Rome, Italy, April 18--20, 2011. Proceedings},
-  year =      {2011},
-  editor =    {Alberto Marchetti{-}Spaccamela and Michael Segal},
-  volume =    {6595},
-  series =    {Lecture Notes in Computer Science},
-  pages =     {151--162},
-  publisher = {Springer},
-  doi =       {10.1007/978-3-642-19754-3_16}
+@inbook{junttila2011,
+  title = {Conflict Propagation and Component Recursion for Canonical Labeling},
+  ISBN = {9783642197543},
+  ISSN = {1611-3349},
+  DOI = {10.1007/978-3-642-19754-3_16},
+  booktitle = {Lecture Notes in Computer Science},
+  publisher = {Springer Berlin Heidelberg},
+  author = {Junttila,  Tommi and Kaski,  Petteri},
+  year = {2011},
+  pages = {151–-162}
 }
 
-@InProceedings{junttila2007,
-  author =     {Tommi Junttila and Petteri Kaski},
-  title =      {Engineering an efficient canonical labeling tool for
-                large and sparse graphs},
-  booktitle =  {Proceedings of the Ninth Workshop on Algorithm Engineering
-                and Experiments and
-                the Fourth Workshop on Analytic Algorithms and
-                Combinatorics},
-  pages =      {135--149},
-  year =       {2007},
-  editor =     {David Applegate and Gerth St{\o}lting Brodal and
-                Daniel Panario and Robert Sedgewick},
-  OPTaddress = {New Orleans, LA},
-  OPTmonth =   {January 6},
-  publisher =  {SIAM},
-  OPTpublisher = {Society for Industrial and Applied Mathematics},
-  doi = {10.1137/1.9781611972870.13}
+@inbook{junttila2007,
+  title = {Engineering an Efficient Canonical Labeling Tool for Large and Sparse Graphs},
+  ISBN = {9781611972870},
+  DOI = {10.1137/1.9781611972870.13},
+  booktitle = {2007 Proceedings of the Ninth Workshop on Algorithm Engineering and Experiments ({ALENEX})},
+  publisher = {Society for Industrial and Applied Mathematics},
+  author = {Junttila,  Tommi and Kaski,  Petteri},
+  year = {2007},
+  month = jan,
+  pages = {135–149}
 }
 
 @article{yaverouglu2014,
@@ -157,5 +149,16 @@
   doi={10.1007/s41109-021-00429-9}
 }
 
-@article{moroni2023, doi = {10.21105/joss.05116}, url = {https://doi.org/10.21105/joss.05116}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {83}, pages = {5116}, author = {Claudio Moroni and Pietro Monticone}, title = {MultilayerGraphs.jl: Multilayer Network Science in Julia}, journal = {Journal of Open Source Software} }
+@article{moroni2023,
+  doi = {10.21105/joss.05116},
+  url = {https://doi.org/10.21105/joss.05116},
+  year = 2023,
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {83},
+  pages = {5116},
+  author = {Claudio Moroni and Pietro Monticone},
+  title = {MultilayerGraphs.jl: Multilayer Network Science in {J}ulia},
+  journal = {Journal of Open Source Software}
+}
 
