### 2024-07-23 #6930 @danielskatz
👋 @ercco - my suggested changes are in https://github.com/mnets/pymnet/pull/51  Please merge this, or let me know what you disagree with, then we can proceed.
