### 2024-07-23 #6930 @danielskatz
@ercco - please check this PDF to make sure everything looks ok, including the changes. If so, then:
