### 2024-07-24 99a3b10
adjust fig 1 width for better layout
