### 2024-07-24 #6930 @ercco
The figure placement was a bit awkward in the article pdf, leading to large white spaces at the bottom of pages 2 and 3. I moved and scaled figure 1 slightly to make the layout better.
