### 2024-07-24 #6930 @danielskatz
@ercco - that's fine, this looks better.

I'll continue to await your actions on the bullets [above](https://github.com/openjournals/joss-reviews/issues/6930#issuecomment-2245398454).
