+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pymnet"
-version = "0.3.0"
+version = "1.0.0"
 description = "Multilayer network analysis library for Python"
 authors = [{name="Mikko Kivelä", email="mikko.kivela@iki.fi"}]
 requires-python = ">=3.8"
