### 2024-07-24 #6930 @ercco
@danielskatz Here are the archived and tagged versions of pymnet - we bumped the version number to 1.0.0:

- Zenodo DOI: https://doi.org/10.5281/zenodo.12806499
- Tagged release on GitHub: https://github.com/mnets/pymnet/releases/tag/v1.0.0
- PyPI release of v1.0.0: https://pypi.org/project/pymnet/1.0.0/
