### 2024-07-24 #6930 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
