### 2024-07-24 #6930 @danielskatz
Congratulations to @ercco (Tarmo Nurmi) and co-authors on your publication!!

And thanks to @ClaudMor, @pitmonticone and @nwlandry for reviewing!
JOSS depends on volunteers and wouldn't be successful without your work
