### 2024-07-25 #6930 @ercco
On behalf of all of us, a huge thank you to @ClaudMor, @pitmonticone, @nwlandry, and @danielskatz! All of your comments and suggestions really made the library and the paper much better, thank you for your work. This was a very smooth process overall.
